Enum webkit2gtk::ContextMenuAction

#[non_exhaustive]pub enum ContextMenuAction {
    NoAction,
    OpenLink,
    OpenLinkInNewWindow,
    DownloadLinkToDisk,
    CopyLinkToClipboard,
    OpenImageInNewWindow,
    DownloadImageToDisk,
    CopyImageToClipboard,
    CopyImageUrlToClipboard,
    OpenFrameInNewWindow,
    GoBack,
    GoForward,
    Stop,
    Reload,
    Copy,
    Cut,
    Paste,
    Delete,
    SelectAll,
    InputMethods,
    Unicode,
    SpellingGuess,
    NoGuessesFound,
    IgnoreSpelling,
    LearnSpelling,
    IgnoreGrammar,
    FontMenu,
    Bold,
    Italic,
    Underline,
    Outline,
    InspectElement,
    OpenVideoInNewWindow,
    OpenAudioInNewWindow,
    CopyVideoLinkToClipboard,
    CopyAudioLinkToClipboard,
    ToggleMediaControls,
    ToggleMediaLoop,
    EnterVideoFullscreen,
    MediaPlay,
    MediaPause,
    MediaMute,
    DownloadVideoToDisk,
    DownloadAudioToDisk,
    InsertEmoji,
    PasteAsPlainText,
    Custom,
    // some variants omitted
}

Enum values used to denote the stock actions for ContextMenuItems

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NoAction

No action, used by separator menu items.

OpenLink

Open current link.

OpenLinkInNewWindow

Open current link in a new window.

DownloadLinkToDisk

Download link destination.

CopyLinkToClipboard

Copy link location to the clipboard.

OpenImageInNewWindow

Open current image in a new window.

DownloadImageToDisk

Download current image.

CopyImageToClipboard

Copy current image to the clipboard.

CopyImageUrlToClipboard

Copy current image location to the clipboard.

OpenFrameInNewWindow

Open current frame in a new window.

GoBack

Load the previous history item.

GoForward

Load the next history item.

Stop

Stop any ongoing loading operation.

Reload

Reload the contents of current view.

Copy

Copy current selection the clipboard.

Cut

Cut current selection to the clipboard.

Paste

Paste clipboard contents.

Delete

Delete current selection.

SelectAll

Select all text.

InputMethods

Input methods menu.

Unicode

Unicode menu.

SpellingGuess

A proposed replacement for a misspelled word.

NoGuessesFound

An indicator that spellchecking found no proposed replacements.

IgnoreSpelling

Causes the spellchecker to ignore the word for this session.

LearnSpelling

Causes the spellchecker to add the word to the dictionary.

IgnoreGrammar

Ignore grammar.

FontMenu

Font options menu.

Bold

Bold.

Italic

Italic.

Underline

Underline.

Outline

Outline.

InspectElement

Open current element in the inspector.

OpenVideoInNewWindow

Open current video element in a new window.

OpenAudioInNewWindow

Open current audio element in a new window.

CopyVideoLinkToClipboard

Copy video link location in to the clipboard.

CopyAudioLinkToClipboard

Copy audio link location in to the clipboard.

ToggleMediaControls

Enable or disable media controls.

ToggleMediaLoop

Enable or disable media loop.

EnterVideoFullscreen

Show current video element in fullscreen mode.

MediaPlay

Play current media element.

MediaPause

Pause current media element.

MediaMute

Mute current media element.

DownloadVideoToDisk

Download video to disk. Since 2.2

DownloadAudioToDisk

Download audio to disk. Since 2.2

InsertEmoji

Insert an emoji. Since 2.26

PasteAsPlainText

Paste clipboard contents as plain text. Since 2.30

Custom

Custom action defined by applications.