Struct webkit2gtk::Error

#[repr(transparent)]
pub struct Error { /* private fields */ }

A generic error capable of representing various error domains (types).

Implementations

impl Error

pub fn as_ptr(&self) -> *mut GError

Return the inner pointer to the underlying C value.

impl Error

pub fn new<T>(error: T, message: &str) -> Errorwhere T: ErrorDomain,

Creates an error with supplied error enum variant and message.

pub fn is<T>(&self) -> boolwhere T: ErrorDomain,

Checks if the error domain matches T.

pub fn domain(&self) -> Quark

Returns the error domain quark

pub fn matches<T>(&self, err: T) -> boolwhere T: ErrorDomain,

Checks if the error matches the specified domain and error code.

pub fn kind<T>(&self) -> Option<T>where T: ErrorDomain,

Tries to convert to a specific error enum.

Returns Some if the error belongs to the enum’s error domain and None otherwise.

Examples

if let Some(file_error) = error.kind::<FileError>() {
    match file_error {
        FileError::Exist => ...
        FileError::Isdir => ...
        ...
    }
}

pub fn message(&self) -> &str

Returns the error message

Most of the time you can simply print the error since it implements the Display trait, but you can use this method if you need to have the message as a &str.