Trait webkit2gtk::AuthenticationRequestExt

pub trait AuthenticationRequestExt: 'static {
    fn can_save_credentials(&self) -> bool;

    fn cancel(&self);

    fn get_host(&self) -> Option<GString>;

    fn get_port(&self) -> u32;

    fn get_proposed_credential(&self) -> Option<Credential>;

    fn get_realm(&self) -> Option<GString>;

    fn get_scheme(&self) -> AuthenticationScheme;

    fn is_for_proxy(&self) -> bool;

    fn is_retry(&self) -> bool;

    fn connect_cancelled<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;
}

Trait containing all AuthenticationRequest methods.

Feature: v2_2

Implementors

AuthenticationRequest

Required methods

fn can_save_credentials(&self) -> bool

Determine whether the authentication method associated with this AuthenticationRequest should allow the storage of credentials. This will return false if WebKit doesn't support credential storing, if private browsing is enabled, or if persistent credential storage has been disabled in WebsiteDataManager, unless credentials saving has been explicitly enabled with AuthenticationRequestExt::set_can_save_credentials.

Feature: v2_2

Returns

true if WebKit can store credentials or false otherwise.

fn cancel(&self)

Cancel the authentication challenge. This will also cancel the page loading and result in a WebView::load-failed signal with a NetworkError of type NetworkError::Cancelled being emitted.

Feature: v2_2

fn get_host(&self) -> Option<GString>

Get the host that this authentication challenge is applicable to.

Feature: v2_2

Returns

The host of self.

fn get_port(&self) -> u32

Get the port that this authentication challenge is applicable to.

Feature: v2_2

Returns

The port of self.

fn get_proposed_credential(&self) -> Option<Credential>

Get the Credential of the proposed authentication challenge that was stored from a previous session. The client can use this directly for authentication or construct their own Credential.

Feature: v2_2

Returns

A Credential encapsulating credential details or None if there is no stored credential.

fn get_realm(&self) -> Option<GString>

Get the realm that this authentication challenge is applicable to.

Feature: v2_2

Returns

The realm of self.

fn get_scheme(&self) -> AuthenticationScheme

Get the authentication scheme of the authentication challenge.

Feature: v2_2

Returns

The AuthenticationScheme of self.

fn is_for_proxy(&self) -> bool

Determine whether the authentication challenge is associated with a proxy server rather than an "origin" server.

Feature: v2_2

Returns

true if authentication is for a proxy or false otherwise.

fn is_retry(&self) -> bool

Determine whether this this is a first attempt or a retry for this authentication challenge.

Feature: v2_2

Returns

true if authentication attempt is a retry or false otherwise.

fn connect_cancelled<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted when the user authentication request is cancelled. It allows the application to dismiss its authentication dialog in case of page load failure for example.

Feature: v2_2

Loading content...