Trait webkit2gtk::BackForwardListExt

pub trait BackForwardListExt: 'static {
    fn get_back_item(&self) -> Option<BackForwardListItem>;

    fn get_back_list(&self) -> Vec<BackForwardListItem>;

    fn get_back_list_with_limit(&self, limit: u32) -> Vec<BackForwardListItem>;

    fn get_current_item(&self) -> Option<BackForwardListItem>;

    fn get_forward_item(&self) -> Option<BackForwardListItem>;

    fn get_forward_list(&self) -> Vec<BackForwardListItem>;

    fn get_forward_list_with_limit(
        &self, 
        limit: u32
    ) -> Vec<BackForwardListItem>;

    fn get_length(&self) -> u32;

    fn get_nth_item(&self, index: i32) -> Option<BackForwardListItem>;
}

Trait containing all BackForwardList methods.

Implementors

BackForwardList

Required methods

fn get_back_item(&self) -> Option<BackForwardListItem>

Returns the item that precedes the current item.

Returns

the BackForwardListItem preceding the current item or None.

fn get_back_list(&self) -> Vec<BackForwardListItem>

Returns

a glib::List of items preceding the current item.

fn get_back_list_with_limit(&self, limit: u32) -> Vec<BackForwardListItem>

limit

the number of items to retrieve

Returns

a glib::List of items preceding the current item limited by limit.

fn get_current_item(&self) -> Option<BackForwardListItem>

Returns the current item in self.

Returns

a BackForwardListItem or None if self is empty.

fn get_forward_item(&self) -> Option<BackForwardListItem>

Returns the item that follows the current item.

Returns

the BackForwardListItem following the current item or None.

fn get_forward_list(&self) -> Vec<BackForwardListItem>

Returns

a glib::List of items following the current item.

fn get_forward_list_with_limit(&self, limit: u32) -> Vec<BackForwardListItem>

limit

the number of items to retrieve

Returns

a glib::List of items following the current item limited by limit.

fn get_length(&self) -> u32

Returns

the length of self.

fn get_nth_item(&self, index: i32) -> Option<BackForwardListItem>

Returns the item at a given index relative to the current item.

index

the index of the item

Returns

the BackForwardListItem located at the specified index relative to the current item or None.

Loading content...