Trait webkit2gtk::ColorChooserRequestExt

pub trait ColorChooserRequestExt: 'static {
    fn cancel(&self);

    fn finish(&self);

    fn get_element_rectangle(&self) -> Rectangle;

    fn get_rgba(&self) -> RGBA;

    fn set_rgba(&self, rgba: &RGBA);

    fn get_property_rgba(&self) -> Option<RGBA>;

    fn set_property_rgba(&self, rgba: Option<&RGBA>);

    fn connect_finished<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn connect_property_rgba_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all ColorChooserRequest methods.

Feature: v2_8

Implementors

ColorChooserRequest

Required methods

fn cancel(&self)

Cancels self and the input element changes to use the initial color it has before the request started. The signal ColorChooserRequest::finished is emitted to notify that the request has finished.

Feature: v2_8

fn finish(&self)

Finishes self and the input element keeps the current value of ColorChooserRequest:rgba. The signal ColorChooserRequest::finished is emitted to notify that the request has finished.

Feature: v2_8

fn get_element_rectangle(&self) -> Rectangle

Gets the bounding box of the color input element.

Feature: v2_8

rect

a gdk::Rectangle to fill in with the element area

fn get_rgba(&self) -> RGBA

Gets the current gdk::RGBA color of self

Feature: v2_8

rgba

a gdk::RGBA to fill in with the current color.

fn set_rgba(&self, rgba: &RGBA)

Sets the current gdk::RGBA color of self

Feature: v2_8

rgba

a pointer gdk::RGBA

fn get_property_rgba(&self) -> Option<RGBA>

fn set_property_rgba(&self, rgba: Option<&RGBA>)

fn connect_finished<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

Emitted when the request finishes. This signal can be emitted because the user completed the request calling ColorChooserRequestExt::finish, or cancelled it with ColorChooserRequestExt::cancel or because the color input element is removed from the DOM.

Feature: v2_8

fn connect_property_rgba_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...