Trait webkit2gtk::ContextMenuExt

pub trait ContextMenuExt: 'static {
    fn append<P: IsA<ContextMenuItem>>(&self, item: &P);

    fn first(&self) -> Option<ContextMenuItem>;

    fn get_item_at_position(&self, position: u32) -> Option<ContextMenuItem>;

    fn get_items(&self) -> Vec<ContextMenuItem>;

    fn get_n_items(&self) -> u32;

    fn get_user_data(&self) -> Option<Variant>;

    fn insert<P: IsA<ContextMenuItem>>(&self, item: &P, position: i32);

    fn last(&self) -> Option<ContextMenuItem>;

    fn move_item<P: IsA<ContextMenuItem>>(&self, item: &P, position: i32);

    fn prepend<P: IsA<ContextMenuItem>>(&self, item: &P);

    fn remove<P: IsA<ContextMenuItem>>(&self, item: &P);

    fn remove_all(&self);

    fn set_user_data(&self, user_data: &Variant);
}

Trait containing all ContextMenu methods.

Implementors

ContextMenu

Required methods

fn append<P: IsA<ContextMenuItem>>(&self, item: &P)

Adds item at the end of the self.

item

the ContextMenuItem to add

fn first(&self) -> Option<ContextMenuItem>

Gets the first item in the self.

Returns

the first ContextMenuItem of self, or None if the ContextMenu is empty.

fn get_item_at_position(&self, position: u32) -> Option<ContextMenuItem>

Gets the item at the given position in the self.

position

the position of the item, counting from 0

Returns

the ContextMenuItem at position position in self, or None if the position is off the end of the self.

fn get_items(&self) -> Vec<ContextMenuItem>

Returns the item list of self.

Returns

a glib::List of ContextMenuItems

fn get_n_items(&self) -> u32

Gets the length of the self.

Returns

the number of ContextMenuItems in self

fn get_user_data(&self) -> Option<Variant>

Gets the user data of self. This function can be used from the UI Process to get user data previously set from the Web Process with ContextMenuExt::set_user_data.

Feature: v2_8

Returns

the user data of self, or None if self doesn't have user data

fn insert<P: IsA<ContextMenuItem>>(&self, item: &P, position: i32)

Inserts item into the self at the given position. If position is negative, or is larger than the number of items in the ContextMenu, the item is added on to the end of the self. The first position is 0.

item

the ContextMenuItem to add

position

the position to insert the item

fn last(&self) -> Option<ContextMenuItem>

Gets the last item in the self.

Returns

the last ContextMenuItem of self, or None if the ContextMenu is empty.

fn move_item<P: IsA<ContextMenuItem>>(&self, item: &P, position: i32)

Moves item to the given position in the self. If position is negative, or is larger than the number of items in the ContextMenu, the item is added on to the end of the self. The first position is 0.

item

the ContextMenuItem to add

position

the new position to move the item

fn prepend<P: IsA<ContextMenuItem>>(&self, item: &P)

Adds item at the beginning of the self.

item

the ContextMenuItem to add

fn remove<P: IsA<ContextMenuItem>>(&self, item: &P)

Removes item from the self. See also ContextMenuExt::remove_all to remove all items.

item

the ContextMenuItem to remove

fn remove_all(&self)

Removes all items of the self.

fn set_user_data(&self, user_data: &Variant)

Sets user data to self. This function can be used from a Web Process extension to set user data that can be retrieved from the UI Process using ContextMenuExt::get_user_data. If the user_data glib::Variant is floating, it is consumed.

Feature: v2_8

user_data

a glib::Variant

Loading content...