Trait webkit2gtk::ContextMenuItemExt

pub trait ContextMenuItemExt: 'static {
    fn get_gaction(&self) -> Option<Action>;

    fn get_stock_action(&self) -> ContextMenuAction;

    fn get_submenu(&self) -> Option<ContextMenu>;

    fn is_separator(&self) -> bool;

    fn set_submenu<P: IsA<ContextMenu>>(&self, submenu: Option<&P>);
}

Trait containing all ContextMenuItem methods.

Implementors

ContextMenuItem

Required methods

fn get_gaction(&self) -> Option<Action>

Gets the action associated to self as a gio::Action.

Feature: v2_18

Returns

the gio::Action associated to the ContextMenuItem, or None if self is a separator.

fn get_stock_action(&self) -> ContextMenuAction

Gets the ContextMenuAction of self. If the ContextMenuItem was not created for a stock action ContextMenuAction::Custom will be returned. If the ContextMenuItem is a separator ContextMenuAction::NoAction will be returned.

Returns

the ContextMenuAction of self

fn get_submenu(&self) -> Option<ContextMenu>

Gets the submenu of self.

Returns

the ContextMenu representing the submenu of self or None if self doesn't have a submenu.

fn is_separator(&self) -> bool

Checks whether self is a separator.

Returns

true is self is a separator or false otherwise

fn set_submenu<P: IsA<ContextMenu>>(&self, submenu: Option<&P>)

Sets or replaces the self submenu. If submenu is None the current submenu of self is removed.

submenu

a ContextMenu

Loading content...