Trait webkit2gtk::EditorStateExt

pub trait EditorStateExt: 'static {
    fn get_typing_attributes(&self) -> u32;

    fn is_copy_available(&self) -> bool;

    fn is_cut_available(&self) -> bool;

    fn is_paste_available(&self) -> bool;

    fn is_redo_available(&self) -> bool;

    fn is_undo_available(&self) -> bool;

    fn connect_property_typing_attributes_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all EditorState methods.

Feature: v2_10

Implementors

EditorState

Required methods

fn get_typing_attributes(&self) -> u32

Gets the typing attributes at the current cursor position. If there is a selection, this returns the typing attributes of the selected text. Note that in case of a selection, typing attributes are considered active only when they are present throughout the selection.

Feature: v2_10

Returns

a bitmask of EditorTypingAttributes flags

fn is_copy_available(&self) -> bool

Gets whether a copy command can be issued.

Feature: v2_20

Returns

true if copy is currently available

fn is_cut_available(&self) -> bool

Gets whether a cut command can be issued.

Feature: v2_20

Returns

true if cut is currently available

fn is_paste_available(&self) -> bool

Gets whether a paste command can be issued.

Feature: v2_20

Returns

true if paste is currently available

fn is_redo_available(&self) -> bool

Gets whether a redo command can be issued.

Feature: v2_20

Returns

true if redo is currently available

fn is_undo_available(&self) -> bool

Gets whether an undo command can be issued.

Feature: v2_20

Returns

true if undo is currently available

fn connect_property_typing_attributes_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...