Trait webkit2gtk::FaviconDatabaseExt

pub trait FaviconDatabaseExt: 'static {
    fn clear(&self);

    fn get_favicon<P: IsA<Cancellable>, Q: FnOnce(Result<Surface, Error>) + Send + 'static>(
        &self, 
        page_uri: &str, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn get_favicon_future(
        &self, 
        page_uri: &str
    ) -> Pin<Box_<dyn Future<Output = Result<Surface, Error>> + 'static>>;

    fn get_favicon_uri(&self, page_uri: &str) -> Option<GString>;

    fn connect_favicon_changed<F: Fn(&Self, &str, &str) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all FaviconDatabase methods.

Implementors

FaviconDatabase

Required methods

fn clear(&self)

Clears all icons from the database.

fn get_favicon<P: IsA<Cancellable>, Q: FnOnce(Result<Surface, Error>) + Send + 'static>(
&self,
page_uri: &str,
cancellable: Option<&P>,
callback: Q
)

Asynchronously obtains a cairo::Surface of the favicon for the given page URI. It returns the cached icon if it's in the database asynchronously waiting for the icon to be read from the database.

This is an asynchronous method. When the operation is finished, callback will be invoked. You can then call FaviconDatabaseExt::get_favicon_finish to get the result of the operation.

You must call WebContextExt::set_favicon_database_directory for the WebContext associated with this FaviconDatabase before attempting to use this function; otherwise, FaviconDatabaseExt::get_favicon_finish will return FaviconDatabaseError::NotInitialized.

page_uri

URI of the page for which we want to retrieve the favicon

cancellable

A gio::Cancellable or None.

callback

A GAsyncReadyCallback to call when the request is satisfied or None if you don't care about the result.

user_data

The data to pass to callback.

fn get_favicon_future(
&self,
page_uri: &str
) -> Pin<Box_<dyn Future<Output = Result<Surface, Error>> + 'static>>

fn get_favicon_uri(&self, page_uri: &str) -> Option<GString>

Obtains the URI of the favicon for the given page_uri.

page_uri

URI of the page containing the icon

Returns

a newly allocated URI for the favicon, or None if the database doesn't have a favicon for page_uri.

fn connect_favicon_changed<F: Fn(&Self, &str, &str) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when the favicon URI of page_uri has been changed to favicon_uri in the database. You can connect to this signal and call FaviconDatabaseExt::get_favicon to get the favicon. If you are interested in the favicon of a WebView it's easier to use the WebView:favicon property. See WebViewExt::get_favicon for more details.

page_uri

the URI of the Web page containing the icon

favicon_uri

the URI of the favicon

Loading content...