Trait webkit2gtk::FileChooserRequestExt

pub trait FileChooserRequestExt: 'static {
    fn cancel(&self);

    fn get_mime_types(&self) -> Vec<GString>;

    fn get_mime_types_filter(&self) -> Option<FileFilter>;

    fn get_select_multiple(&self) -> bool;

    fn get_selected_files(&self) -> Vec<GString>;

    fn select_files(&self, files: &[&str]);

    fn get_property_filter(&self) -> Option<FileFilter>;

    fn connect_property_filter_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_mime_types_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_select_multiple_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_selected_files_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all FileChooserRequest methods.

Implementors

FileChooserRequest

Required methods

fn cancel(&self)

Ask WebKit to cancel the request. It's important to do this in case no selection has been made in the client, otherwise the request won't be properly completed and the browser will keep the request pending forever, which might cause the browser to hang.

fn get_mime_types(&self) -> Vec<GString>

Get the list of MIME types the file chooser dialog should handle, in the format specified in RFC 2046 for "media types". Its contents depend on the value of the 'accept' attribute for HTML input elements. This function should normally be called before presenting the file chooser dialog to the user, to decide whether to allow the user to select multiple files at once or only one.

Returns

a None-terminated array of strings if a list of accepted MIME types is defined or None otherwise, meaning that any MIME type should be accepted. This array and its contents are owned by WebKit and should not be modified or freed.

fn get_mime_types_filter(&self) -> Option<FileFilter>

Get the filter currently associated with the request, ready to be used by gtk::FileChooser. This function should normally be called before presenting the file chooser dialog to the user, to decide whether to apply a filter so the user would not be allowed to select files with other MIME types.

See FileChooserRequestExt::get_mime_types if you are interested in getting the list of accepted MIME types.

Returns

a gtk::FileFilter if a list of accepted MIME types is defined or None otherwise. The returned object is owned by WebKit should not be modified or freed.

fn get_select_multiple(&self) -> bool

Determine whether the file chooser associated to this FileChooserRequest should allow selecting multiple files, which depends on the HTML input element having a 'multiple' attribute defined.

Returns

true if the file chooser should allow selecting multiple files or false otherwise.

fn get_selected_files(&self) -> Vec<GString>

Get the list of selected files currently associated to the request. Initially, the return value of this method contains any files selected in previous file chooser requests for this HTML input element. Once webkit_file_chooser_request_select_files, the value will reflect whatever files are given.

This function should normally be called only before presenting the file chooser dialog to the user, to decide whether to perform some extra action, like pre-selecting the files from a previous request.

Returns

a None-terminated array of strings if there are selected files associated with the request or None otherwise. This array and its contents are owned by WebKit and should not be modified or freed.

fn select_files(&self, files: &[&str])

Ask WebKit to select local files for upload and complete the request.

files

a None-terminated array of strings, containing paths to local files.

fn get_property_filter(&self) -> Option<FileFilter>

The filter currently associated with the request. See FileChooserRequestExt::get_mime_types_filter for more details.

fn connect_property_filter_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_mime_types_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_select_multiple_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_selected_files_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...