Trait webkit2gtk::HitTestResultExt

pub trait HitTestResultExt: 'static {
    fn context_is_editable(&self) -> bool;

    fn context_is_image(&self) -> bool;

    fn context_is_link(&self) -> bool;

    fn context_is_media(&self) -> bool;

    fn context_is_scrollbar(&self) -> bool;

    fn context_is_selection(&self) -> bool;

    fn get_context(&self) -> u32;

    fn get_image_uri(&self) -> Option<GString>;

    fn get_link_label(&self) -> Option<GString>;

    fn get_link_title(&self) -> Option<GString>;

    fn get_link_uri(&self) -> Option<GString>;

    fn get_media_uri(&self) -> Option<GString>;
}

Trait containing all HitTestResult methods.

Implementors

HitTestResult

Required methods

fn context_is_editable(&self) -> bool

Gets whether HitTestResultContext::Editable flag is present in HitTestResult:context.

Returns

true if there's an editable element at the coordinates of the self, or false otherwise

fn context_is_image(&self) -> bool

Gets whether HitTestResultContext::Image flag is present in HitTestResult:context.

Returns

true if there's an image element in the coordinates of the Hit Test, or false otherwise

fn context_is_link(&self) -> bool

Gets whether HitTestResultContext::Link flag is present in HitTestResult:context.

Returns

true if there's a link element in the coordinates of the Hit Test, or false otherwise

fn context_is_media(&self) -> bool

Gets whether HitTestResultContext::Media flag is present in HitTestResult:context.

Returns

true if there's a media element in the coordinates of the Hit Test, or false otherwise

fn context_is_scrollbar(&self) -> bool

Gets whether HitTestResultContext::Scrollbar flag is present in HitTestResult:context.

Returns

true if there's a scrollbar element at the coordinates of the self, or false otherwise

fn context_is_selection(&self) -> bool

Gets whether HitTestResultContext::Selection flag is present in HitTestResult:context.

Feature: v2_8

Returns

true if there's a selected element at the coordinates of the self, or false otherwise

fn get_context(&self) -> u32

Gets the value of the HitTestResult:context property.

Returns

a bitmask of HitTestResultContext flags

fn get_image_uri(&self) -> Option<GString>

Gets the value of the HitTestResult:image-uri property.

Returns

the URI of the image element in the coordinates of the Hit Test, or None if there isn't an image element in self context

fn get_link_label(&self) -> Option<GString>

Gets the value of the HitTestResult:link-label property.

Returns

the label of the link element in the coordinates of the Hit Test, or None if there isn't a link element in self context or the link element doesn't have a label

fn get_link_title(&self) -> Option<GString>

Gets the value of the HitTestResult:link-title property.

Returns

the title of the link element in the coordinates of the Hit Test, or None if there isn't a link element in self context or the link element doesn't have a title

fn get_link_uri(&self) -> Option<GString>

Gets the value of the HitTestResult:link-uri property.

Returns

the URI of the link element in the coordinates of the Hit Test, or None if there isn't a link element in self context

fn get_media_uri(&self) -> Option<GString>

Gets the value of the HitTestResult:media-uri property.

Returns

the URI of the media element in the coordinates of the Hit Test, or None if there isn't a media element in self context

Loading content...