Trait webkit2gtk::InputMethodContextExt

pub trait InputMethodContextExt: 'static {
    fn filter_key_event(&self, key_event: &mut EventKey) -> bool;

    fn get_input_hints(&self) -> InputHints;

    fn get_input_purpose(&self) -> InputPurpose;

    fn notify_cursor_area(&self, x: i32, y: i32, width: i32, height: i32);

    fn notify_focus_in(&self);

    fn notify_focus_out(&self);

    fn notify_surrounding(
        &self, 
        text: &str, 
        cursor_index: u32, 
        selection_index: u32
    );

    fn reset(&self);

    fn set_enable_preedit(&self, enabled: bool);

    fn set_input_hints(&self, hints: InputHints);

    fn set_input_purpose(&self, purpose: InputPurpose);

    fn get_property_input_hints(&self) -> InputHints;

    fn set_property_input_hints(&self, input_hints: InputHints);

    fn get_property_input_purpose(&self) -> InputPurpose;

    fn set_property_input_purpose(&self, input_purpose: InputPurpose);

    fn connect_committed<F: Fn(&Self, &str) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_delete_surrounding<F: Fn(&Self, i32, u32) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_preedit_changed<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_preedit_finished<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_preedit_started<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_input_hints_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_input_purpose_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all InputMethodContext methods.

Feature: v2_28

Implementors

InputMethodContext

Required methods

fn filter_key_event(&self, key_event: &mut EventKey) -> bool

Allow key_event to be handled by the input method. If true is returned, then no further processing should be done for the key event.

Feature: v2_28

key_event

the key event to filter

Returns

true if the key event was handled, or false otherwise

fn get_input_hints(&self) -> InputHints

Get the value of the InputMethodContext:input-hints property.

Feature: v2_28

Returns

the InputHints of the input associated with self

fn get_input_purpose(&self) -> InputPurpose

Get the value of the InputMethodContext:input-purpose property.

Feature: v2_28

Returns

the InputPurpose of the input associated with self

fn notify_cursor_area(&self, x: i32, y: i32, width: i32, height: i32)

Notify self that cursor area changed in input associated.

Feature: v2_28

x

the x coordinate of cursor location

y

the y coordinate of cursor location

width

the width of cursor area

height

the height of cursor area

fn notify_focus_in(&self)

Notify self that input associated has gained focus.

Feature: v2_28

fn notify_focus_out(&self)

Notify self that input associated has lost focus.

Feature: v2_28

fn notify_surrounding(
&self,
text: &str,
cursor_index: u32,
selection_index: u32
)

Notify self that the context surrounding the cursor has changed. If there's no selection selection_index is the same as cursor_index.

Feature: v2_28

text

text surrounding the insertion point

length

the length of text, or -1 if text is nul-terminated

cursor_index

the byte index of the insertion cursor within text.

selection_index

the byte index of the selection cursor within text.

fn reset(&self)

Reset the self. This will typically cause the input to clear the preedit state.

Feature: v2_28

fn set_enable_preedit(&self, enabled: bool)

Set whether self should enable preedit to display feedback.

Feature: v2_28

enabled

whether to enable preedit

fn set_input_hints(&self, hints: InputHints)

fn set_input_purpose(&self, purpose: InputPurpose)

Set the value of the InputMethodContext:input-purpose property.

Feature: v2_28

purpose

a InputPurpose

fn get_property_input_hints(&self) -> InputHints

fn set_property_input_hints(&self, input_hints: InputHints)

fn get_property_input_purpose(&self) -> InputPurpose

fn set_property_input_purpose(&self, input_purpose: InputPurpose)

fn connect_committed<F: Fn(&Self, &str) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a complete input sequence has been entered by the user. This can be a single character immediately after a key press or the final result of preediting.

Feature: v2_28

text

the string result

fn connect_delete_surrounding<F: Fn(&Self, i32, u32) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when the input method wants to delete the context surrounding the cursor. If offset is a negative value, it means a position before the cursor.

Feature: v2_28

offset

the character offset from the cursor position of the text to be deleted.

n_chars

the number of characters to be deleted

fn connect_preedit_changed<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted whenever the preedit sequence currently being entered has changed. It is also emitted at the end of a preedit sequence, in which case InputMethodContextExt::get_preedit returns the empty string.

Feature: v2_28

fn connect_preedit_finished<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a preediting sequence has been completed or canceled.

Feature: v2_28

fn connect_preedit_started<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted when a new preediting sequence starts.

Feature: v2_28

fn connect_property_input_hints_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_input_purpose_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...