Trait webkit2gtk::InstallMissingMediaPluginsPermissionRequestExt

pub trait InstallMissingMediaPluginsPermissionRequestExt: 'static {
    fn get_description(&self) -> Option<GString>;
}

Trait containing all InstallMissingMediaPluginsPermissionRequest methods.

Feature: v2_10

Implementors

InstallMissingMediaPluginsPermissionRequest

Required methods

fn get_description(&self) -> Option<GString>

Gets the description about the missing plugins provided by the media backend when a media couldn't be played.

Feature: v2_10

Returns

a string with the description provided by the media backend.

Loading content...