Trait webkit2gtk::NavigationPolicyDecisionExt

pub trait NavigationPolicyDecisionExt: 'static {
    fn get_frame_name(&self) -> Option<GString>;

    fn get_modifiers(&self) -> u32;

    fn get_mouse_button(&self) -> u32;

    fn get_navigation_action(&self) -> Option<NavigationAction>;

    fn get_navigation_type(&self) -> NavigationType;

    fn get_request(&self) -> Option<URIRequest>;

    fn connect_property_frame_name_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_modifiers_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_mouse_button_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_navigation_action_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_navigation_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_request_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all NavigationPolicyDecision methods.

Implementors

NavigationPolicyDecision

Required methods

fn get_frame_name(&self) -> Option<GString>

Gets the value of the NavigationPolicyDecision:frame-name property.

Returns

The name of the new frame this navigation action targets or None

fn get_modifiers(&self) -> u32

👎 Deprecated

Gets the value of the NavigationPolicyDecision:modifiers property.

Deprecated since 2.6

Use NavigationPolicyDecisionExt::get_navigation_action instead.

Returns

The modifiers active if this decision was triggered by a mouse event

fn get_mouse_button(&self) -> u32

👎 Deprecated

Gets the value of the NavigationPolicyDecision:mouse-button property.

Deprecated since 2.6

Use NavigationPolicyDecisionExt::get_navigation_action instead.

Returns

The mouse button used if this decision was triggered by a mouse event or 0 otherwise

fn get_navigation_action(&self) -> Option<NavigationAction>

Gets the value of the NavigationPolicyDecision:navigation-action property.

Feature: v2_6

Returns

The NavigationAction triggering this policy decision.

fn get_navigation_type(&self) -> NavigationType

👎 Deprecated

Gets the value of the NavigationPolicyDecision:navigation-type property.

Deprecated since 2.6

Use NavigationPolicyDecisionExt::get_navigation_action instead.

Returns

The type of navigation triggering this policy decision.

fn get_request(&self) -> Option<URIRequest>

👎 Deprecated

Gets the value of the NavigationPolicyDecision:request property.

Deprecated since 2.6

Use NavigationPolicyDecisionExt::get_navigation_action instead.

Returns

The URI request that is associated with this navigation

fn connect_property_frame_name_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_modifiers_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

👎 Deprecated

fn connect_property_mouse_button_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

👎 Deprecated

fn connect_property_navigation_action_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_navigation_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

👎 Deprecated

fn connect_property_request_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

👎 Deprecated

Loading content...