Trait webkit2gtk::PrintCustomWidgetExt

pub trait PrintCustomWidgetExt: 'static {
    fn get_title(&self) -> Option<GString>;

    fn get_widget(&self) -> Option<Widget>;

    fn connect_apply<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn connect_update<F: Fn(&Self, &PageSetup, &PrintSettings) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all PrintCustomWidget methods.

Feature: v2_16

Implementors

PrintCustomWidget

Required methods

fn get_title(&self) -> Option<GString>

Return the value of PrintCustomWidget:title property for the given self object.

Feature: v2_16

Returns

Title of the self.

fn get_widget(&self) -> Option<Widget>

Return the value of PrintCustomWidget:widget property for the given self object. The returned value will always be valid if called from PrintCustomWidget::apply or PrintCustomWidget::update callbacks, but it will be None if called after the PrintCustomWidget::apply signal is emitted.

Feature: v2_16

Returns

a gtk::Widget.

fn connect_apply<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

Emitted right before the printing will start. You should read the information from the widget and update the content based on it if necessary. The widget is not guaranteed to be valid at a later time.

Feature: v2_16

fn connect_update<F: Fn(&Self, &PageSetup, &PrintSettings) + 'static>(
&self,
f: F
) -> SignalHandlerId

Emitted after change of selected printer in the dialog. The actual page setup and print settings are available and the custom widget can actualize itself according to their values.

Feature: v2_16

page_setup

actual page setup

print_settings

actual print settings

Loading content...