Trait webkit2gtk::ResponsePolicyDecisionExt

pub trait ResponsePolicyDecisionExt: 'static {
    fn get_request(&self) -> Option<URIRequest>;

    fn get_response(&self) -> Option<URIResponse>;

    fn is_mime_type_supported(&self) -> bool;

    fn connect_property_request_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_response_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all ResponsePolicyDecision methods.

Implementors

ResponsePolicyDecision

Required methods

fn get_request(&self) -> Option<URIRequest>

Return the URIRequest associated with the response decision. Modifications to the returned object are <emphasis>not</emphasis> taken into account when the request is sent over the network, and is intended only to aid in evaluating whether a response decision should be taken or not. To modify requests before they are sent over the network the WebKitPage::send-request signal can be used instead.

Returns

The URI request that is associated with this policy decision.

fn get_response(&self) -> Option<URIResponse>

Gets the value of the ResponsePolicyDecision:response property.

Returns

The URI response that is associated with this policy decision.

fn is_mime_type_supported(&self) -> bool

Gets whether the MIME type of the response can be displayed in the WebView that triggered this policy decision request. See also WebViewExt::can_show_mime_type.

Feature: v2_4

Returns

true if the MIME type of the response is supported or false otherwise

fn connect_property_request_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_response_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...