Trait webkit2gtk::SecurityManagerExt

pub trait SecurityManagerExt: 'static {
    fn register_uri_scheme_as_cors_enabled(&self, scheme: &str);

    fn register_uri_scheme_as_display_isolated(&self, scheme: &str);

    fn register_uri_scheme_as_empty_document(&self, scheme: &str);

    fn register_uri_scheme_as_local(&self, scheme: &str);

    fn register_uri_scheme_as_no_access(&self, scheme: &str);

    fn register_uri_scheme_as_secure(&self, scheme: &str);

    fn uri_scheme_is_cors_enabled(&self, scheme: &str) -> bool;

    fn uri_scheme_is_display_isolated(&self, scheme: &str) -> bool;

    fn uri_scheme_is_empty_document(&self, scheme: &str) -> bool;

    fn uri_scheme_is_local(&self, scheme: &str) -> bool;

    fn uri_scheme_is_no_access(&self, scheme: &str) -> bool;

    fn uri_scheme_is_secure(&self, scheme: &str) -> bool;
}

Trait containing all SecurityManager methods.

Implementors

SecurityManager

Required methods

fn register_uri_scheme_as_cors_enabled(&self, scheme: &str)

Register scheme as a CORS (Cross-origin resource sharing) enabled scheme. This means that CORS requests are allowed. See W3C CORS specification http://www.w3.org/TR/cors/.

scheme

a URI scheme

fn register_uri_scheme_as_display_isolated(&self, scheme: &str)

Register scheme as a display isolated scheme. This means that pages cannot display these URIs unless they are from the same scheme.

scheme

a URI scheme

fn register_uri_scheme_as_empty_document(&self, scheme: &str)

Register scheme as an empty document scheme. This means that they are allowed to commit synchronously.

scheme

a URI scheme

fn register_uri_scheme_as_local(&self, scheme: &str)

Register scheme as a local scheme. This means that other non-local pages cannot link to or access URIs of this scheme.

scheme

a URI scheme

fn register_uri_scheme_as_no_access(&self, scheme: &str)

Register scheme as a no-access scheme. This means that pages loaded with this URI scheme cannot access pages loaded with any other URI scheme.

scheme

a URI scheme

fn register_uri_scheme_as_secure(&self, scheme: &str)

Register scheme as a secure scheme. This means that mixed content warnings won't be generated for this scheme when included by an HTTPS page.

scheme

a URI scheme

fn uri_scheme_is_cors_enabled(&self, scheme: &str) -> bool

Whether scheme is considered as a CORS enabled scheme. See also SecurityManagerExt::register_uri_scheme_as_cors_enabled.

scheme

a URI scheme

Returns

true if scheme is a CORS enabled scheme or false otherwise.

fn uri_scheme_is_display_isolated(&self, scheme: &str) -> bool

Whether scheme is considered as a display isolated scheme. See also SecurityManagerExt::register_uri_scheme_as_display_isolated.

scheme

a URI scheme

Returns

true if scheme is a display isolated scheme or false otherwise.

fn uri_scheme_is_empty_document(&self, scheme: &str) -> bool

Whether scheme is considered as an empty document scheme. See also SecurityManagerExt::register_uri_scheme_as_empty_document.

scheme

a URI scheme

Returns

true if scheme is an empty document scheme or false otherwise.

fn uri_scheme_is_local(&self, scheme: &str) -> bool

Whether scheme is considered as a local scheme. See also SecurityManagerExt::register_uri_scheme_as_local.

scheme

a URI scheme

Returns

true if scheme is a local scheme or false otherwise.

fn uri_scheme_is_no_access(&self, scheme: &str) -> bool

Whether scheme is considered as a no-access scheme. See also SecurityManagerExt::register_uri_scheme_as_no_access.

scheme

a URI scheme

Returns

true if scheme is a no-access scheme or false otherwise.

fn uri_scheme_is_secure(&self, scheme: &str) -> bool

Whether scheme is considered as a secure scheme. See also SecurityManagerExt::register_uri_scheme_as_secure.

scheme

a URI scheme

Returns

true if scheme is a secure scheme or false otherwise.

Loading content...