Trait webkit2gtk::SettingsExt

pub trait SettingsExt: 'static {
    fn get_allow_file_access_from_file_urls(&self) -> bool;

    fn get_allow_modal_dialogs(&self) -> bool;

    fn get_allow_top_navigation_to_data_urls(&self) -> bool;

    fn get_allow_universal_access_from_file_urls(&self) -> bool;

    fn get_auto_load_images(&self) -> bool;

    fn get_cursive_font_family(&self) -> Option<GString>;

    fn get_default_charset(&self) -> Option<GString>;

    fn get_default_font_family(&self) -> Option<GString>;

    fn get_default_font_size(&self) -> u32;

    fn get_default_monospace_font_size(&self) -> u32;

    fn get_draw_compositing_indicators(&self) -> bool;

    fn get_enable_accelerated_2d_canvas(&self) -> bool;

    fn get_enable_back_forward_navigation_gestures(&self) -> bool;

    fn get_enable_caret_browsing(&self) -> bool;

    fn get_enable_developer_extras(&self) -> bool;

    fn get_enable_dns_prefetching(&self) -> bool;

    fn get_enable_encrypted_media(&self) -> bool;

    fn get_enable_frame_flattening(&self) -> bool;

    fn get_enable_fullscreen(&self) -> bool;

    fn get_enable_html5_database(&self) -> bool;

    fn get_enable_html5_local_storage(&self) -> bool;

    fn get_enable_hyperlink_auditing(&self) -> bool;

    fn get_enable_java(&self) -> bool;

    fn get_enable_javascript(&self) -> bool;

    fn get_enable_javascript_markup(&self) -> bool;

    fn get_enable_media(&self) -> bool;

    fn get_enable_media_capabilities(&self) -> bool;

    fn get_enable_media_stream(&self) -> bool;

    fn get_enable_mediasource(&self) -> bool;

    fn get_enable_mock_capture_devices(&self) -> bool;

    fn get_enable_offline_web_application_cache(&self) -> bool;

    fn get_enable_page_cache(&self) -> bool;

    fn get_enable_plugins(&self) -> bool;

    fn get_enable_private_browsing(&self) -> bool;

    fn get_enable_resizable_text_areas(&self) -> bool;

    fn get_enable_site_specific_quirks(&self) -> bool;

    fn get_enable_smooth_scrolling(&self) -> bool;

    fn get_enable_spatial_navigation(&self) -> bool;

    fn get_enable_tabs_to_links(&self) -> bool;

    fn get_enable_webaudio(&self) -> bool;

    fn get_enable_webgl(&self) -> bool;

    fn get_enable_write_console_messages_to_stdout(&self) -> bool;

    fn get_enable_xss_auditor(&self) -> bool;

    fn get_fantasy_font_family(&self) -> Option<GString>;

    fn get_hardware_acceleration_policy(&self) -> HardwareAccelerationPolicy;

    fn get_javascript_can_access_clipboard(&self) -> bool;

    fn get_javascript_can_open_windows_automatically(&self) -> bool;

    fn get_load_icons_ignoring_image_load_setting(&self) -> bool;

    fn get_media_playback_allows_inline(&self) -> bool;

    fn get_media_playback_requires_user_gesture(&self) -> bool;

    fn get_minimum_font_size(&self) -> u32;

    fn get_monospace_font_family(&self) -> Option<GString>;

    fn get_pictograph_font_family(&self) -> Option<GString>;

    fn get_print_backgrounds(&self) -> bool;

    fn get_sans_serif_font_family(&self) -> Option<GString>;

    fn get_serif_font_family(&self) -> Option<GString>;

    fn get_user_agent(&self) -> Option<GString>;

    fn get_zoom_text_only(&self) -> bool;

    fn set_allow_file_access_from_file_urls(&self, allowed: bool);

    fn set_allow_modal_dialogs(&self, allowed: bool);

    fn set_allow_top_navigation_to_data_urls(&self, allowed: bool);

    fn set_allow_universal_access_from_file_urls(&self, allowed: bool);

    fn set_auto_load_images(&self, enabled: bool);

    fn set_cursive_font_family(&self, cursive_font_family: &str);

    fn set_default_charset(&self, default_charset: &str);

    fn set_default_font_family(&self, default_font_family: &str);

    fn set_default_font_size(&self, font_size: u32);

    fn set_default_monospace_font_size(&self, font_size: u32);

    fn set_draw_compositing_indicators(&self, enabled: bool);

    fn set_enable_accelerated_2d_canvas(&self, enabled: bool);

    fn set_enable_back_forward_navigation_gestures(&self, enabled: bool);

    fn set_enable_caret_browsing(&self, enabled: bool);

    fn set_enable_developer_extras(&self, enabled: bool);

    fn set_enable_dns_prefetching(&self, enabled: bool);

    fn set_enable_encrypted_media(&self, enabled: bool);

    fn set_enable_frame_flattening(&self, enabled: bool);

    fn set_enable_fullscreen(&self, enabled: bool);

    fn set_enable_html5_database(&self, enabled: bool);

    fn set_enable_html5_local_storage(&self, enabled: bool);

    fn set_enable_hyperlink_auditing(&self, enabled: bool);

    fn set_enable_java(&self, enabled: bool);

    fn set_enable_javascript(&self, enabled: bool);

    fn set_enable_javascript_markup(&self, enabled: bool);

    fn set_enable_media(&self, enabled: bool);

    fn set_enable_media_capabilities(&self, enabled: bool);

    fn set_enable_media_stream(&self, enabled: bool);

    fn set_enable_mediasource(&self, enabled: bool);

    fn set_enable_mock_capture_devices(&self, enabled: bool);

    fn set_enable_offline_web_application_cache(&self, enabled: bool);

    fn set_enable_page_cache(&self, enabled: bool);

    fn set_enable_plugins(&self, enabled: bool);

    fn set_enable_private_browsing(&self, enabled: bool);

    fn set_enable_resizable_text_areas(&self, enabled: bool);

    fn set_enable_site_specific_quirks(&self, enabled: bool);

    fn set_enable_smooth_scrolling(&self, enabled: bool);

    fn set_enable_spatial_navigation(&self, enabled: bool);

    fn set_enable_tabs_to_links(&self, enabled: bool);

    fn set_enable_webaudio(&self, enabled: bool);

    fn set_enable_webgl(&self, enabled: bool);

    fn set_enable_write_console_messages_to_stdout(&self, enabled: bool);

    fn set_enable_xss_auditor(&self, enabled: bool);

    fn set_fantasy_font_family(&self, fantasy_font_family: &str);

    fn set_hardware_acceleration_policy(
        &self, 
        policy: HardwareAccelerationPolicy
    );

    fn set_javascript_can_access_clipboard(&self, enabled: bool);

    fn set_javascript_can_open_windows_automatically(&self, enabled: bool);

    fn set_load_icons_ignoring_image_load_setting(&self, enabled: bool);

    fn set_media_playback_allows_inline(&self, enabled: bool);

    fn set_media_playback_requires_user_gesture(&self, enabled: bool);

    fn set_minimum_font_size(&self, font_size: u32);

    fn set_monospace_font_family(&self, monospace_font_family: &str);

    fn set_pictograph_font_family(&self, pictograph_font_family: &str);

    fn set_print_backgrounds(&self, print_backgrounds: bool);

    fn set_sans_serif_font_family(&self, sans_serif_font_family: &str);

    fn set_serif_font_family(&self, serif_font_family: &str);

    fn set_user_agent(&self, user_agent: Option<&str>);

    fn set_user_agent_with_application_details(
        &self, 
        application_name: Option<&str>, 
        application_version: Option<&str>
    );

    fn set_zoom_text_only(&self, zoom_text_only: bool);

    fn connect_property_allow_file_access_from_file_urls_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_allow_modal_dialogs_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_allow_top_navigation_to_data_urls_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_allow_universal_access_from_file_urls_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_auto_load_images_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_cursive_font_family_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_default_charset_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_default_font_family_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_default_font_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_default_monospace_font_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_draw_compositing_indicators_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_accelerated_2d_canvas_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_back_forward_navigation_gestures_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_caret_browsing_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_developer_extras_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_dns_prefetching_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_encrypted_media_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_frame_flattening_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_fullscreen_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_html5_database_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_html5_local_storage_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_hyperlink_auditing_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_java_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_javascript_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_javascript_markup_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_media_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_media_capabilities_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_media_stream_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_mediasource_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_mock_capture_devices_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_offline_web_application_cache_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_page_cache_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_plugins_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_private_browsing_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_resizable_text_areas_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_site_specific_quirks_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_smooth_scrolling_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_spatial_navigation_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_tabs_to_links_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_webaudio_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_webgl_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_write_console_messages_to_stdout_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_xss_auditor_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_fantasy_font_family_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_hardware_acceleration_policy_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_javascript_can_access_clipboard_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_javascript_can_open_windows_automatically_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_load_icons_ignoring_image_load_setting_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_media_playback_allows_inline_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_media_playback_requires_user_gesture_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_minimum_font_size_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_monospace_font_family_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_pictograph_font_family_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_print_backgrounds_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_sans_serif_font_family_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_serif_font_family_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_user_agent_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_zoom_text_only_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Settings methods.

Implementors

Settings

Required methods

fn get_allow_file_access_from_file_urls(&self) -> bool

Get the Settings:allow-file-access-from-file-urls property.

Feature: v2_10

Returns

true If file access from file URLs is allowed or false otherwise.

fn get_allow_modal_dialogs(&self) -> bool

Get the Settings:allow-modal-dialogs property.

Returns

true if it's allowed to create and run modal dialogs or false otherwise.

fn get_allow_top_navigation_to_data_urls(&self) -> bool

Get the Settings:allow-top-navigation-to-data-urls property.

Feature: v2_28

Returns

true If navigation to data URLs from the top frame is allowed or false
otherwise.

fn get_allow_universal_access_from_file_urls(&self) -> bool

Get the Settings:allow-universal-access-from-file-urls property.

Feature: v2_14

Returns

true If universal access from file URLs is allowed or false otherwise.

fn get_auto_load_images(&self) -> bool

Get the Settings:auto-load-images property.

Returns

true If auto loading of images is enabled or false otherwise.

fn get_cursive_font_family(&self) -> Option<GString>

Gets the Settings:cursive-font-family property.

Returns

The default font family used to display content marked with cursive font.

fn get_default_charset(&self) -> Option<GString>

Gets the Settings:default-charset property.

Returns

Default charset.

fn get_default_font_family(&self) -> Option<GString>

fn get_default_font_size(&self) -> u32

Gets the Settings:default-font-size property.

Returns

The default font size, in pixels.

fn get_default_monospace_font_size(&self) -> u32

Gets the Settings:default-monospace-font-size property.

Returns

Default monospace font size, in pixels.

fn get_draw_compositing_indicators(&self) -> bool

Get the Settings:draw-compositing-indicators property.

Returns

true If compositing borders are drawn or false otherwise.

fn get_enable_accelerated_2d_canvas(&self) -> bool

Get the Settings:enable-accelerated-2d-canvas property.

Feature: v2_2

Returns

true if accelerated 2D canvas is enabled or false otherwise.

fn get_enable_back_forward_navigation_gestures(&self) -> bool

Get the Settings:enable-back-forward-navigation-gestures property.

Feature: v2_24

Returns

true if horizontal swipe gesture will trigger back-forward navigaiton or false otherwise.

fn get_enable_caret_browsing(&self) -> bool

Get the Settings:enable-caret-browsing property.

Returns

true If caret browsing is enabled or false otherwise.

fn get_enable_developer_extras(&self) -> bool

Get the Settings:enable-developer-extras property.

Returns

true If developer extras is enabled or false otherwise.

fn get_enable_dns_prefetching(&self) -> bool

Get the Settings:enable-dns-prefetching property.

Returns

true If DNS prefetching is enabled or false otherwise.

fn get_enable_encrypted_media(&self) -> bool

Get the Settings:enable-encrypted-media property.

Feature: v2_20

Returns

true if EncryptedMedia support is enabled or false otherwise.

fn get_enable_frame_flattening(&self) -> bool

Get the Settings:enable-frame-flattening property.

Returns

true If frame flattening is enabled or false otherwise.

fn get_enable_fullscreen(&self) -> bool

Get the Settings:enable-fullscreen property.

Returns

true If fullscreen support is enabled or false otherwise.

fn get_enable_html5_database(&self) -> bool

Get the Settings:enable-html5-database property.

Returns

true if IndexedDB support is enabled or false otherwise.

fn get_enable_html5_local_storage(&self) -> bool

Get the Settings:enable-html5-local-storage property.

Returns

true If HTML5 local storage support is enabled or false otherwise.

fn get_enable_hyperlink_auditing(&self) -> bool

Get the Settings:enable-hyperlink-auditing property.

Returns

true If hyper link auditing is enabled or false otherwise.

fn get_enable_java(&self) -> bool

Get the Settings:enable-java property.

Returns

true If Java is enabled or false otherwise.

fn get_enable_javascript(&self) -> bool

Get the Settings:enable-javascript property.

Returns

true If JavaScript is enabled or false otherwise.

fn get_enable_javascript_markup(&self) -> bool

Get the Settings:enable-javascript-markup property.

Feature: v2_24

Returns

true if JavaScript markup is enabled or false otherwise.

fn get_enable_media(&self) -> bool

Get the Settings:enable-media property.

Feature: v2_26

Returns

true if media support is enabled or false otherwise.

fn get_enable_media_capabilities(&self) -> bool

Get the Settings:enable-media-capabilities property.

Feature: v2_22

Returns

true if MediaCapabilities support is enabled or false otherwise.

fn get_enable_media_stream(&self) -> bool

Get the Settings:enable-media-stream property.

Feature: v2_4

Returns

true If mediastream support is enabled or false otherwise.

fn get_enable_mediasource(&self) -> bool

Get the Settings:enable-mediasource property.

Feature: v2_4

Returns

true If MediaSource support is enabled or false otherwise.

fn get_enable_mock_capture_devices(&self) -> bool

Get the Settings:enable-mock-capture-devices property.

Feature: v2_24

Returns

true If mock capture devices is enabled or false otherwise.

fn get_enable_offline_web_application_cache(&self) -> bool

Get the Settings:enable-offline-web-application-cache property.

Returns

true If HTML5 offline web application cache support is enabled or false otherwise.

fn get_enable_page_cache(&self) -> bool

Get the Settings:enable-page-cache property.

Returns

true if page cache enabled or false otherwise.

fn get_enable_plugins(&self) -> bool

Get the Settings:enable-plugins property.

Returns

true If plugins are enabled or false otherwise.

fn get_enable_private_browsing(&self) -> bool

👎 Deprecated

Get the Settings:enable-private-browsing property.

Deprecated since 2.16

Use WebView:is-ephemeral or WebContext:is-ephemeral instead.

Returns

true If private browsing is enabled or false otherwise.

fn get_enable_resizable_text_areas(&self) -> bool

Get the Settings:enable-resizable-text-areas property.

Returns

true If text areas can be resized or false otherwise.

fn get_enable_site_specific_quirks(&self) -> bool

Get the Settings:enable-site-specific-quirks property.

Returns

true if site specific quirks are enabled or false otherwise.

fn get_enable_smooth_scrolling(&self) -> bool

Get the Settings:enable-smooth-scrolling property.

Returns

true if smooth scrolling is enabled or false otherwise.

fn get_enable_spatial_navigation(&self) -> bool

Get the Settings:enable-spatial-navigation property.

Feature: v2_2

Returns

true If HTML5 spatial navigation support is enabled or false otherwise.

fn get_enable_tabs_to_links(&self) -> bool

Get the Settings:enable-tabs-to-links property.

Returns

true If tabs to link is enabled or false otherwise.

fn get_enable_webaudio(&self) -> bool

Get the Settings:enable-webaudio property.

Returns

true If webaudio support is enabled or false otherwise.

fn get_enable_webgl(&self) -> bool

Get the Settings:enable-webgl property.

Returns

true If WebGL support is enabled or false otherwise.

fn get_enable_write_console_messages_to_stdout(&self) -> bool

Get the Settings:enable-write-console-messages-to-stdout property.

Feature: v2_2

Returns

true if writing console messages to stdout is enabled or false otherwise.

fn get_enable_xss_auditor(&self) -> bool

Get the Settings:enable-xss-auditor property.

Returns

true If XSS auditing is enabled or false otherwise.

fn get_fantasy_font_family(&self) -> Option<GString>

Gets the Settings:fantasy-font-family property.

Returns

The default font family used to display content marked with fantasy font.

fn get_hardware_acceleration_policy(&self) -> HardwareAccelerationPolicy

Get the Settings:hardware-acceleration-policy property.

Feature: v2_16

Returns

a HardwareAccelerationPolicy

fn get_javascript_can_access_clipboard(&self) -> bool

Get the Settings:javascript-can-access-clipboard property.

Returns

true If javascript-can-access-clipboard is enabled or false otherwise.

fn get_javascript_can_open_windows_automatically(&self) -> bool

Get the Settings:javascript-can-open-windows-automatically property.

Returns

true If JavaScript can open window automatically or false otherwise.

fn get_load_icons_ignoring_image_load_setting(&self) -> bool

Get the Settings:load-icons-ignoring-image-load-setting property.

Returns

true If site icon can be loaded irrespective of image loading preference or false otherwise.

fn get_media_playback_allows_inline(&self) -> bool

Get the Settings:media-playback-allows-inline property.

Returns

true If inline playback is allowed for media or false if only fullscreen playback is allowed.

fn get_media_playback_requires_user_gesture(&self) -> bool

Get the Settings:media-playback-requires-user-gesture property.

Returns

true If an user gesture is needed to play or load media or false if no user gesture is needed.

fn get_minimum_font_size(&self) -> u32

Gets the Settings:minimum-font-size property.

Returns

Minimum font size, in pixels.

fn get_monospace_font_family(&self) -> Option<GString>

Gets the Settings:monospace-font-family property.

Returns

Default font family used to display content marked with monospace font.

fn get_pictograph_font_family(&self) -> Option<GString>

Gets the Settings:pictograph-font-family property.

Returns

The default font family used to display content marked with pictograph font.

fn get_print_backgrounds(&self) -> bool

Get the Settings:print-backgrounds property.

Returns

true If background images should be printed or false otherwise.

fn get_sans_serif_font_family(&self) -> Option<GString>

Gets the Settings:sans-serif-font-family property.

Returns

The default font family used to display content marked with sans-serif font.

fn get_serif_font_family(&self) -> Option<GString>

Gets the Settings:serif-font-family property.

Returns

The default font family used to display content marked with serif font.

fn get_user_agent(&self) -> Option<GString>

Get the Settings:user-agent property.

Returns

The current value of the user-agent property.

fn get_zoom_text_only(&self) -> bool

Get the Settings:zoom-text-only property.

Returns

true If zoom level of the view should only affect the text or false if all view contents should be scaled.

fn set_allow_file_access_from_file_urls(&self, allowed: bool)

Set the Settings:allow-file-access-from-file-urls property.

Feature: v2_10

allowed

Value to be set

fn set_allow_modal_dialogs(&self, allowed: bool)

Set the Settings:allow-modal-dialogs property.

allowed

Value to be set

fn set_allow_top_navigation_to_data_urls(&self, allowed: bool)

Set the Settings:allow-top-navigation-to-data-urls property.

Feature: v2_28

allowed

Value to be set

fn set_allow_universal_access_from_file_urls(&self, allowed: bool)

Set the Settings:allow-universal-access-from-file-urls property.

Feature: v2_14

allowed

Value to be set

fn set_auto_load_images(&self, enabled: bool)

Set the Settings:auto-load-images property.

enabled

Value to be set

fn set_cursive_font_family(&self, cursive_font_family: &str)

Set the Settings:cursive-font-family property.

cursive_font_family

the new default cursive font family

fn set_default_charset(&self, default_charset: &str)

Set the Settings:default-charset property.

default_charset

default charset to be set

fn set_default_font_family(&self, default_font_family: &str)

Set the Settings:default-font-family property.

default_font_family

the new default font family

fn set_default_font_size(&self, font_size: u32)

Set the Settings:default-font-size property.

font_size

default font size to be set in pixels

fn set_default_monospace_font_size(&self, font_size: u32)

Set the Settings:default-monospace-font-size property.

font_size

default monospace font size to be set in pixels

fn set_draw_compositing_indicators(&self, enabled: bool)

Set the Settings:draw-compositing-indicators property.

enabled

Value to be set

fn set_enable_accelerated_2d_canvas(&self, enabled: bool)

Set the Settings:enable-accelerated-2d-canvas property.

Feature: v2_2

enabled

Value to be set

fn set_enable_back_forward_navigation_gestures(&self, enabled: bool)

Set the Settings:enable-back-forward-navigation-gestures property.

Feature: v2_24

enabled

value to be set

fn set_enable_caret_browsing(&self, enabled: bool)

Set the Settings:enable-caret-browsing property.

enabled

Value to be set

fn set_enable_developer_extras(&self, enabled: bool)

Set the Settings:enable-developer-extras property.

enabled

Value to be set

fn set_enable_dns_prefetching(&self, enabled: bool)

Set the Settings:enable-dns-prefetching property.

enabled

Value to be set

fn set_enable_encrypted_media(&self, enabled: bool)

Set the Settings:enable-encrypted-media property.

Feature: v2_20

enabled

Value to be set

fn set_enable_frame_flattening(&self, enabled: bool)

Set the Settings:enable-frame-flattening property.

enabled

Value to be set

fn set_enable_fullscreen(&self, enabled: bool)

Set the Settings:enable-fullscreen property.

enabled

Value to be set

fn set_enable_html5_database(&self, enabled: bool)

Set the Settings:enable-html5-database property.

enabled

Value to be set

fn set_enable_html5_local_storage(&self, enabled: bool)

Set the Settings:enable-html5-local-storage property.

enabled

Value to be set

fn set_enable_hyperlink_auditing(&self, enabled: bool)

Set the Settings:enable-hyperlink-auditing property.

enabled

Value to be set

fn set_enable_java(&self, enabled: bool)

Set the Settings:enable-java property.

enabled

Value to be set

fn set_enable_javascript(&self, enabled: bool)

Set the Settings:enable-javascript property.

enabled

Value to be set

fn set_enable_javascript_markup(&self, enabled: bool)

Set the Settings:enable-javascript-markup property.

Feature: v2_24

enabled

Value to be set

fn set_enable_media(&self, enabled: bool)

Set the Settings:enable-media property.

Feature: v2_26

enabled

Value to be set

fn set_enable_media_capabilities(&self, enabled: bool)

Set the Settings:enable-media-capabilities property.

Feature: v2_22

enabled

Value to be set

fn set_enable_media_stream(&self, enabled: bool)

Set the Settings:enable-media-stream property.

Feature: v2_4

enabled

Value to be set

fn set_enable_mediasource(&self, enabled: bool)

Set the Settings:enable-mediasource property.

Feature: v2_4

enabled

Value to be set

fn set_enable_mock_capture_devices(&self, enabled: bool)

Set the Settings:enable-mock-capture-devices property.

Feature: v2_4

enabled

Value to be set

fn set_enable_offline_web_application_cache(&self, enabled: bool)

Set the Settings:enable-offline-web-application-cache property.

enabled

Value to be set

fn set_enable_page_cache(&self, enabled: bool)

Set the Settings:enable-page-cache property.

enabled

Value to be set

fn set_enable_plugins(&self, enabled: bool)

Set the Settings:enable-plugins property.

enabled

Value to be set

fn set_enable_private_browsing(&self, enabled: bool)

👎 Deprecated

Set the Settings:enable-private-browsing property.

Deprecated since 2.16

Use WebView:is-ephemeral or WebContext:is-ephemeral instead.

enabled

Value to be set

fn set_enable_resizable_text_areas(&self, enabled: bool)

Set the Settings:enable-resizable-text-areas property.

enabled

Value to be set

fn set_enable_site_specific_quirks(&self, enabled: bool)

Set the Settings:enable-site-specific-quirks property.

enabled

Value to be set

fn set_enable_smooth_scrolling(&self, enabled: bool)

Set the Settings:enable-smooth-scrolling property.

enabled

Value to be set

fn set_enable_spatial_navigation(&self, enabled: bool)

Set the Settings:enable-spatial-navigation property.

Feature: v2_2

enabled

Value to be set

fn set_enable_tabs_to_links(&self, enabled: bool)

Set the Settings:enable-tabs-to-links property.

enabled

Value to be set

fn set_enable_webaudio(&self, enabled: bool)

Set the Settings:enable-webaudio property.

enabled

Value to be set

fn set_enable_webgl(&self, enabled: bool)

Set the Settings:enable-webgl property.

enabled

Value to be set

fn set_enable_write_console_messages_to_stdout(&self, enabled: bool)

Set the Settings:enable-write-console-messages-to-stdout property.

Feature: v2_2

enabled

Value to be set

fn set_enable_xss_auditor(&self, enabled: bool)

Set the Settings:enable-xss-auditor property.

enabled

Value to be set

fn set_fantasy_font_family(&self, fantasy_font_family: &str)

Set the Settings:fantasy-font-family property.

fantasy_font_family

the new default fantasy font family

fn set_hardware_acceleration_policy(&self, policy: HardwareAccelerationPolicy)

Set the Settings:hardware-acceleration-policy property.

Feature: v2_16

policy

a HardwareAccelerationPolicy

fn set_javascript_can_access_clipboard(&self, enabled: bool)

Set the Settings:javascript-can-access-clipboard property.

enabled

Value to be set

fn set_javascript_can_open_windows_automatically(&self, enabled: bool)

Set the Settings:javascript-can-open-windows-automatically property.

enabled

Value to be set

fn set_load_icons_ignoring_image_load_setting(&self, enabled: bool)

Set the Settings:load-icons-ignoring-image-load-setting property.

enabled

Value to be set

fn set_media_playback_allows_inline(&self, enabled: bool)

Set the Settings:media-playback-allows-inline property.

enabled

Value to be set

fn set_media_playback_requires_user_gesture(&self, enabled: bool)

Set the Settings:media-playback-requires-user-gesture property.

enabled

Value to be set

fn set_minimum_font_size(&self, font_size: u32)

Set the Settings:minimum-font-size property.

font_size

minimum font size to be set in pixels

fn set_monospace_font_family(&self, monospace_font_family: &str)

Set the Settings:monospace-font-family property.

monospace_font_family

the new default monospace font family

fn set_pictograph_font_family(&self, pictograph_font_family: &str)

Set the Settings:pictograph-font-family property.

pictograph_font_family

the new default pictograph font family

fn set_print_backgrounds(&self, print_backgrounds: bool)

Set the Settings:print-backgrounds property.

print_backgrounds

Value to be set

fn set_sans_serif_font_family(&self, sans_serif_font_family: &str)

Set the Settings:sans-serif-font-family property.

sans_serif_font_family

the new default sans-serif font family

fn set_serif_font_family(&self, serif_font_family: &str)

Set the Settings:serif-font-family property.

serif_font_family

the new default serif font family

fn set_user_agent(&self, user_agent: Option<&str>)

Set the Settings:user-agent property.

user_agent

The new custom user agent string or None to use the default user agent

fn set_user_agent_with_application_details(
&self,
application_name: Option<&str>,
application_version: Option<&str>
)

Set the Settings:user-agent property by appending the application details to the default user agent. If no application name or version is given, the default user agent used will be used. If only the version is given, the default engine version is used with the given application name.

application_name

The application name used for the user agent or None to use the default user agent.

application_version

The application version for the user agent or None to user the default version.

fn set_zoom_text_only(&self, zoom_text_only: bool)

Set the Settings:zoom-text-only property.

zoom_text_only

Value to be set

fn connect_property_allow_file_access_from_file_urls_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_allow_modal_dialogs_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_allow_top_navigation_to_data_urls_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_allow_universal_access_from_file_urls_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_auto_load_images_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_cursive_font_family_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_default_charset_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_default_font_family_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_default_font_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_default_monospace_font_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_draw_compositing_indicators_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_accelerated_2d_canvas_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_back_forward_navigation_gestures_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_caret_browsing_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_developer_extras_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_dns_prefetching_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_encrypted_media_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_frame_flattening_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_fullscreen_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_html5_database_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_html5_local_storage_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_hyperlink_auditing_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_java_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_javascript_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_javascript_markup_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_media_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_media_capabilities_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_media_stream_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_mediasource_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_mock_capture_devices_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_offline_web_application_cache_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_page_cache_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_plugins_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_private_browsing_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

👎 Deprecated

fn connect_property_enable_resizable_text_areas_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_site_specific_quirks_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_smooth_scrolling_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_spatial_navigation_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_tabs_to_links_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_webaudio_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_webgl_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_write_console_messages_to_stdout_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_xss_auditor_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_fantasy_font_family_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_hardware_acceleration_policy_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_javascript_can_access_clipboard_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_javascript_can_open_windows_automatically_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_load_icons_ignoring_image_load_setting_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_media_playback_allows_inline_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_media_playback_requires_user_gesture_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_minimum_font_size_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_monospace_font_family_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_pictograph_font_family_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_print_backgrounds_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_sans_serif_font_family_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_serif_font_family_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_user_agent_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_zoom_text_only_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...