Trait webkit2gtk::URIResponseExt

pub trait URIResponseExt: 'static {
    fn get_content_length(&self) -> u64;

    fn get_mime_type(&self) -> Option<GString>;

    fn get_status_code(&self) -> u32;

    fn get_suggested_filename(&self) -> Option<GString>;

    fn get_uri(&self) -> Option<GString>;

    fn connect_property_content_length_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_http_headers_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_mime_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_status_code_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_suggested_filename_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_uri_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all URIResponse methods.

Implementors

URIResponse

Required methods

fn get_content_length(&self) -> u64

Get the expected content length of the URIResponse. It can be 0 if the server provided an incorrect or missing Content-Length.

Returns

the expected content length of self.

fn get_mime_type(&self) -> Option<GString>

Returns

the MIME type of the URIResponse

fn get_status_code(&self) -> u32

Get the status code of the URIResponse as returned by the server. It will normally be a soup::KnownStatusCode, for example soup::Status::Ok, though the server can respond with any unsigned integer.

Returns

the status code of self

fn get_suggested_filename(&self) -> Option<GString>

Get the suggested filename for self, as specified by the 'Content-Disposition' HTTP header, or None if it's not present.

Returns

the suggested filename or None if the 'Content-Disposition' HTTP header is not present.

fn get_uri(&self) -> Option<GString>

Returns

the uri of the URIResponse

fn connect_property_content_length_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_http_headers_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_mime_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_status_code_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_suggested_filename_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_uri_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...