Trait webkit2gtk::URISchemeRequestExt

pub trait URISchemeRequestExt: 'static {
    fn finish<P: IsA<InputStream>>(
        &self, 
        stream: &P, 
        stream_length: i64, 
        content_type: Option<&str>
    );

    fn finish_error(&self, error: &mut Error);

    fn get_path(&self) -> Option<GString>;

    fn get_scheme(&self) -> Option<GString>;

    fn get_uri(&self) -> Option<GString>;

    fn get_web_view(&self) -> Option<WebView>;
}

Trait containing all URISchemeRequest methods.

Implementors

URISchemeRequest

Required methods

fn finish<P: IsA<InputStream>>(
&self,
stream: &P,
stream_length: i64,
content_type: Option<&str>
)

Finish a URISchemeRequest by setting the contents of the request and its mime type.

stream

a gio::InputStream to read the contents of the request

stream_length

the length of the stream or -1 if not known

content_type

the content type of the stream or None if not known

fn finish_error(&self, error: &mut Error)

Finish a URISchemeRequest with a glib::Error.

Feature: v2_2

error

a glib::Error that will be passed to the WebView

fn get_path(&self) -> Option<GString>

Get the URI path of self

Returns

the URI path of self

fn get_scheme(&self) -> Option<GString>

Get the URI scheme of self

Returns

the URI scheme of self

fn get_uri(&self) -> Option<GString>

Get the URI of self

Returns

the full URI of self

fn get_web_view(&self) -> Option<WebView>

Get the WebView that initiated the request.

Returns

the WebView that initiated self.

Loading content...