Trait webkit2gtk::UserContentManagerExt

pub trait UserContentManagerExt: 'static {
    fn add_script(&self, script: &UserScript);

    fn add_style_sheet(&self, stylesheet: &UserStyleSheet);

    fn register_script_message_handler(&self, name: &str) -> bool;

    fn register_script_message_handler_in_world(
        &self, 
        name: &str, 
        world_name: &str
    ) -> bool;

    fn remove_all_filters(&self);

    fn remove_all_scripts(&self);

    fn remove_all_style_sheets(&self);

    fn remove_filter_by_id(&self, filter_id: &str);

    fn unregister_script_message_handler(&self, name: &str);

    fn unregister_script_message_handler_in_world(
        &self, 
        name: &str, 
        world_name: &str
    );

    fn connect_script_message_received<F: Fn(&Self, &JavascriptResult) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all UserContentManager methods.

Feature: v2_6

Implementors

UserContentManager

Required methods

fn add_script(&self, script: &UserScript)

Adds a UserScript to the given UserContentManager. The same UserScript can be reused with multiple UserContentManager instances.

Feature: v2_6

script

A UserScript

fn add_style_sheet(&self, stylesheet: &UserStyleSheet)

Adds a UserStyleSheet to the given UserContentManager. The same UserStyleSheet can be reused with multiple UserContentManager instances.

Feature: v2_6

stylesheet

A UserStyleSheet

fn register_script_message_handler(&self, name: &str) -> bool

Registers a new user script message handler. After it is registered, scripts can use window.webkit.messageHandlers.&lt;name&gt;.postMessage(value) to send messages. Those messages are received by connecting handlers to the UserContentManager::script-message-received signal. The handler name is used as the detail of the signal. To avoid race conditions between registering the handler name, and starting to receive the signals, it is recommended to connect to the signal before registering the handler name:

<informalexample>``<programlisting> WebKitWebView *view = webkit_web_view_new (); WebKitUserContentManager *manager = webkit_web_view_get_user_content_manager (); g_signal_connect (manager, "script-message-received::foobar", G_CALLBACK (handle_script_message), NULL); webkit_user_content_manager_register_script_message_handler (manager, "foobar"); </programlisting>``</informalexample>

Registering a script message handler will fail if the requested name has been already registered before.

Feature: v2_8

name

Name of the script message channel

Returns

true if message handler was registered successfully, or false otherwise.

fn register_script_message_handler_in_world(
&self,
name: &str,
world_name: &str
) -> bool

Registers a new user script message handler in script world with name world_name. See UserContentManagerExt::register_script_message_handler for full description.

Registering a script message handler will fail if the requested name has been already registered before.

Feature: v2_22

name

Name of the script message channel

world_name

the name of a WebKitScriptWorld

Returns

true if message handler was registered successfully, or false otherwise.

fn remove_all_filters(&self)

Removes all content filters from the given UserContentManager.

Feature: v2_24

fn remove_all_scripts(&self)

Removes all user scripts from the given UserContentManager

Feature: v2_6

fn remove_all_style_sheets(&self)

Removes all user style sheets from the given UserContentManager.

Feature: v2_6

fn remove_filter_by_id(&self, filter_id: &str)

Removes a filter from the given UserContentManager given the identifier of a UserContentFilter as returned by UserContentFilter::get_identifier.

Feature: v2_26

filter_id

Filter identifier

fn unregister_script_message_handler(&self, name: &str)

Unregisters a previously registered message handler.

Note that this does not disconnect handlers for the UserContentManager::script-message-received signal; they will be kept connected, but the signal will not be emitted unless the handler name is registered again.

See also UserContentManagerExt::register_script_message_handler.

Feature: v2_8

name

Name of the script message channel

fn unregister_script_message_handler_in_world(
&self,
name: &str,
world_name: &str
)

Unregisters a previously registered message handler in script world with name world_name.

Note that this does not disconnect handlers for the UserContentManager::script-message-received signal; they will be kept connected, but the signal will not be emitted unless the handler name is registered again.

See also UserContentManagerExt::register_script_message_handler_in_world.

Feature: v2_22

name

Name of the script message channel

world_name

the name of a WebKitScriptWorld

fn connect_script_message_received<F: Fn(&Self, &JavascriptResult) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when JavaScript in a web view calls <code>window.webkit.messageHandlers.<name>.postMessage()</code>, after registering <code><name></code> using UserContentManagerExt::register_script_message_handler

Feature: v2_8

js_result

the JavascriptResult holding the value received from the JavaScript world.

Loading content...