Trait webkit2gtk::WebContextExt

pub trait WebContextExt: 'static {
    fn add_path_to_sandbox<P: AsRef<Path>>(&self, path: P, read_only: bool);

    fn allow_tls_certificate_for_host<P: IsA<TlsCertificate>>(
        &self, 
        certificate: &P, 
        host: &str
    );

    fn clear_cache(&self);

    fn download_uri(&self, uri: &str) -> Option<Download>;

    fn get_cache_model(&self) -> CacheModel;

    fn get_cookie_manager(&self) -> Option<CookieManager>;

    fn get_favicon_database(&self) -> Option<FaviconDatabase>;

    fn get_favicon_database_directory(&self) -> Option<GString>;

    fn get_geolocation_manager(&self) -> Option<GeolocationManager>;

    fn get_plugins<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<Plugin>, Error>) + Send + 'static>(
        &self, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn get_plugins_future(
        &self
    ) -> Pin<Box_<dyn Future<Output = Result<Vec<Plugin>, Error>> + 'static>>;

    fn get_process_model(&self) -> ProcessModel;

    fn get_sandbox_enabled(&self) -> bool;

    fn get_security_manager(&self) -> Option<SecurityManager>;

    fn get_spell_checking_enabled(&self) -> bool;

    fn get_spell_checking_languages(&self) -> Vec<GString>;

    fn get_tls_errors_policy(&self) -> TLSErrorsPolicy;

    fn get_use_system_appearance_for_scrollbars(&self) -> bool;

    fn get_web_process_count_limit(&self) -> u32;

    fn get_website_data_manager(&self) -> Option<WebsiteDataManager>;

    fn initialize_notification_permissions(
        &self, 
        allowed_origins: &[&SecurityOrigin], 
        disallowed_origins: &[&SecurityOrigin]
    );

    fn is_automation_allowed(&self) -> bool;

    fn is_ephemeral(&self) -> bool;

    fn prefetch_dns(&self, hostname: &str);

    fn register_uri_scheme<P: Fn(&URISchemeRequest) + 'static>(
        &self, 
        scheme: &str, 
        callback: P
    );

    fn send_message_to_all_extensions<P: IsA<UserMessage>>(&self, message: &P);

    fn set_additional_plugins_directory(&self, directory: &str);

    fn set_automation_allowed(&self, allowed: bool);

    fn set_cache_model(&self, cache_model: CacheModel);

    fn set_disk_cache_directory(&self, directory: &str);

    fn set_favicon_database_directory(&self, path: Option<&str>);

    fn set_preferred_languages(&self, languages: &[&str]);

    fn set_process_model(&self, process_model: ProcessModel);

    fn set_sandbox_enabled(&self, enabled: bool);

    fn set_spell_checking_enabled(&self, enabled: bool);

    fn set_spell_checking_languages(&self, languages: &[&str]);

    fn set_tls_errors_policy(&self, policy: TLSErrorsPolicy);

    fn set_use_system_appearance_for_scrollbars(&self, enabled: bool);

    fn set_web_extensions_directory(&self, directory: &str);

    fn set_web_extensions_initialization_user_data(&self, user_data: &Variant);

    fn set_web_process_count_limit(&self, limit: u32);

    fn get_property_local_storage_directory(&self) -> Option<GString>;

    fn get_property_process_swap_on_cross_site_navigation_enabled(&self) -> bool;

    fn connect_automation_started<F: Fn(&Self, &AutomationSession) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_download_started<F: Fn(&Self, &Download) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_initialize_notification_permissions<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_initialize_web_extensions<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_user_message_received<F: Fn(&Self, &UserMessage) -> bool + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_use_system_appearance_for_scrollbars_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all WebContext methods.

Implementors

WebContext

Required methods

fn add_path_to_sandbox<P: AsRef<Path>>(&self, path: P, read_only: bool)

Adds a path to be mounted in the sandbox. path must exist before any web process has been created otherwise it will be silently ignored. It is a fatal error to add paths after a web process has been spawned.

Paths in directories such as /sys, /proc, and /dev or all of / are not valid.

See also WebContextExt::set_sandbox_enabled

Feature: v2_26

path

an absolute path to mount in the sandbox

read_only

if true the path will be read-only

fn allow_tls_certificate_for_host<P: IsA<TlsCertificate>>(
&self,
certificate: &P,
host: &str
)

Ignore further TLS errors on the host for the certificate present in info.

Feature: v2_6

certificate

a gio::TlsCertificate

host

the host for which a certificate is to be allowed

fn clear_cache(&self)

Clears all resources currently cached. See also WebContextExt::set_cache_model.

fn download_uri(&self, uri: &str) -> Option<Download>

Requests downloading of the specified URI string. The download operation will not be associated to any WebView, if you are interested in starting a download from a particular WebView use WebViewExt::download_uri instead.

uri

the URI to download

Returns

a new Download representing the download operation.

fn get_cache_model(&self) -> CacheModel

Returns the current cache model. For more information about this value check the documentation of the function WebContextExt::set_cache_model.

Returns

the current CacheModel

fn get_cookie_manager(&self) -> Option<CookieManager>

Get the CookieManager of the self's WebsiteDataManager.

Returns

the CookieManager of self.

fn get_favicon_database(&self) -> Option<FaviconDatabase>

Get the FaviconDatabase associated with self.

To initialize the database you need to call WebContextExt::set_favicon_database_directory.

Returns

the FaviconDatabase of self.

fn get_favicon_database_directory(&self) -> Option<GString>

Get the directory path being used to store the favicons database for self, or None if WebContextExt::set_favicon_database_directory hasn't been called yet.

This function will always return the same path after having called WebContextExt::set_favicon_database_directory for the first time.

Returns

the path of the directory of the favicons database associated with self, or None.

fn get_geolocation_manager(&self) -> Option<GeolocationManager>

Get the GeolocationManager of self.

Feature: v2_26

Returns

the GeolocationManager of self.

fn get_plugins<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<Plugin>, Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

Asynchronously get the list of installed plugins.

When the operation is finished, callback will be called. You can then call WebContextExt::get_plugins_finish to get the result of the operation.

cancellable

a gio::Cancellable or None to ignore

callback

a GAsyncReadyCallback to call when the request is satisfied

user_data

the data to pass to callback function

fn get_plugins_future(
&self
) -> Pin<Box_<dyn Future<Output = Result<Vec<Plugin>, Error>> + 'static>>

fn get_process_model(&self) -> ProcessModel

Returns the current process model. For more information about this value see WebContextExt::set_process_model.

Feature: v2_4

Returns

the current ProcessModel

fn get_sandbox_enabled(&self) -> bool

Get whether sandboxing is currently enabled.

Feature: v2_26

Returns

true if sandboxing is enabled, or false otherwise.

fn get_security_manager(&self) -> Option<SecurityManager>

Get the SecurityManager of self.

Returns

the SecurityManager of self.

fn get_spell_checking_enabled(&self) -> bool

Get whether spell checking feature is currently enabled.

Returns

true If spell checking is enabled, or false otherwise.

fn get_spell_checking_languages(&self) -> Vec<GString>

Get the the list of spell checking languages associated with self, or None if no languages have been previously set.

See WebContextExt::set_spell_checking_languages for more details on the format of the languages in the list.

Returns

A None-terminated array of languages if available, or None otherwise.

fn get_tls_errors_policy(&self) -> TLSErrorsPolicy

Get the TLS errors policy of self

Returns

a TLSErrorsPolicy

fn get_use_system_appearance_for_scrollbars(&self) -> bool

Get the WebContext:use-system-appearance-for-scrollbars property.

Feature: v2_30

Returns

true if scrollbars are rendering using the system appearance, or false otherwise

fn get_web_process_count_limit(&self) -> u32

👎 Deprecated

Gets the maximum number of web processes that can be created at the same time for the self.

This function is now deprecated and always returns 0 (no limit). See also WebContextExt::set_web_process_count_limit.

Feature: v2_10

Deprecated since 2.26

Returns

the maximum limit of web processes, or 0 if there isn't a limit.

fn get_website_data_manager(&self) -> Option<WebsiteDataManager>

Get the WebsiteDataManager of self.

Feature: v2_10

Returns

a WebsiteDataManager

fn initialize_notification_permissions(
&self,
allowed_origins: &[&SecurityOrigin],
disallowed_origins: &[&SecurityOrigin]
)

Sets initial desktop notification permissions for the self. allowed_origins and disallowed_origins must each be glib::List of SecurityOrigin objects representing origins that will, respectively, either always or never have permission to show desktop notifications. No NotificationPermissionRequest will ever be generated for any of the security origins represented in allowed_origins or disallowed_origins. This function is necessary because some webpages proactively check whether they have permission to display notifications without ever creating a permission request.

This function only affects web processes that have not already been created. The best time to call it is when handling WebContext::initialize-notification-permissions so as to ensure that new web processes receive the most recent set of permissions.

Feature: v2_16

allowed_origins

a glib::List of security origins

disallowed_origins

a glib::List of security origins

fn is_automation_allowed(&self) -> bool

Get whether automation is allowed in self. See also WebContextExt::set_automation_allowed.

Feature: v2_18

Returns

true if automation is allowed or false otherwise.

fn is_ephemeral(&self) -> bool

Get whether a WebContext is ephemeral.

Feature: v2_16

Returns

true if self is ephemeral or false otherwise.

fn prefetch_dns(&self, hostname: &str)

Resolve the domain name of the given hostname in advance, so that if a URI of hostname is requested the load will be performed more quickly.

hostname

a hostname to be resolved

fn register_uri_scheme<P: Fn(&URISchemeRequest) + 'static>(
&self,
scheme: &str,
callback: P
)

Register scheme in self, so that when an URI request with scheme is made in the WebContext, the WebKitURISchemeRequestCallback registered will be called with a URISchemeRequest. It is possible to handle URI scheme requests asynchronously, by calling gobject::ObjectExt::ref on the URISchemeRequest and calling URISchemeRequestExt::finish later when the data of the request is available or URISchemeRequestExt::finish_error in case of error.

<informalexample>``<programlisting> static void about_uri_scheme_request_cb (WebKitURISchemeRequest *request, gpointer user_data) { GInputStream *stream; gsize stream_length; const gchar *path;

path = webkit_uri_scheme_request_get_path (request); if (!g_strcmp0 (path, "plugins")) { /* Create a GInputStream with the contents of plugins about page, and set its length to stream_length / } else if (!g_strcmp0 (path, "memory")) { / Create a GInputStream with the contents of memory about page, and set its length to stream_length / } else if (!g_strcmp0 (path, "applications")) { / Create a GInputStream with the contents of applications about page, and set its length to stream_length */ } else if (!g_strcmp0 (path, "example")) { gchar *contents;

contents = g_strdup_printf ("<html><body><p>Example about page</p></body></html>"); stream_length = strlen (contents); stream = g_memory_input_stream_new_from_data (contents, stream_length, g_free); } else { GError *error;

error = g_error_new (ABOUT_HANDLER_ERROR, ABOUT_HANDLER_ERROR_INVALID, "Invalid about:s page.", path); webkit_uri_scheme_request_finish_error (request, error); g_error_free (error); return; } webkit_uri_scheme_request_finish (request, stream, stream_length, "text/html"); g_object_unref (stream); } </programlisting>``</informalexample>

scheme

the network scheme to register

callback

a WebKitURISchemeRequestCallback

user_data

data to pass to callback function

user_data_destroy_func

destroy notify for user_data

fn send_message_to_all_extensions<P: IsA<UserMessage>>(&self, message: &P)

Send message to all WebKitWebExtensions associated to self. If message is floating, it's consumed.

Feature: v2_28

message

a UserMessage

fn set_additional_plugins_directory(&self, directory: &str)

Set an additional directory where WebKit will look for plugins.

directory

the directory to add

fn set_automation_allowed(&self, allowed: bool)

Set whether automation is allowed in self. When automation is enabled the browser could be controlled by another process by requesting an automation session. When a new automation session is requested the signal WebContext::automation-started is emitted. Automation is disabled by default, so you need to explicitly call this method passing true to enable it.

Note that only one WebContext can have automation enabled, so this will do nothing if there's another WebContext with automation already enabled.

Feature: v2_18

allowed

value to set

fn set_cache_model(&self, cache_model: CacheModel)

Specifies a usage model for WebViews, which WebKit will use to determine its caching behavior. All web views follow the cache model. This cache model determines the RAM and disk space to use for caching previously viewed content .

Research indicates that users tend to browse within clusters of documents that hold resources in common, and to revisit previously visited documents. WebKit and the frameworks below it include built-in caches that take advantage of these patterns, substantially improving document load speed in browsing situations. The WebKit cache model controls the behaviors of all of these caches, including various WebCore caches.

Browsers can improve document load speed substantially by specifying CacheModel::WebBrowser. Applications without a browsing interface can reduce memory usage substantially by specifying CacheModel::DocumentViewer. The default value is CacheModel::WebBrowser.

cache_model

a CacheModel

fn set_disk_cache_directory(&self, directory: &str)

👎 Deprecated

Set the directory where disk cache files will be stored This method must be called before loading anything in this context, otherwise it will not have any effect.

Note that this method overrides the directory set in the WebsiteDataManager, but it doesn't change the value returned by WebsiteDataManagerExt::get_disk_cache_directory since the WebsiteDataManager is immutable.

Deprecated since 2.10

Use WebContext::new_with_website_data_manager instead.

directory

the directory to set

fn set_favicon_database_directory(&self, path: Option<&str>)

Set the directory path to be used to store the favicons database for self on disk. Passing None as path means using the default directory for the platform (see g_get_user_cache_dir).

Calling this method also means enabling the favicons database for its use from the applications, so that's why it's expected to be called only once. Further calls for the same instance of WebContext won't cause any effect.

path

an absolute path to the icon database directory or None to use the defaults

fn set_preferred_languages(&self, languages: &[&str])

Set the list of preferred languages, sorted from most desirable to least desirable. The list will be used to build the "Accept-Language" header that will be included in the network requests started by the WebContext.

languages

a None-terminated list of language identifiers

fn set_process_model(&self, process_model: ProcessModel)

Specifies a process model for WebViews, which WebKit will use to determine how auxiliary processes are handled.

ProcessModel::MultipleSecondaryProcesses will use one process per view most of the time, while still allowing for web views to share a process when needed (for example when different views interact with each other). Using this model, when a process hangs or crashes, only the WebViews using it stop working, while the rest of the WebViews in the application will still function normally.

ProcessModel::SharedSecondaryProcess is deprecated since 2.26, using it has no effect for security reasons.

This method must be called before any web process has been created, as early as possible in your application. Calling it later will make your application crash.

Feature: v2_4

process_model

a ProcessModel

fn set_sandbox_enabled(&self, enabled: bool)

Set whether WebKit subprocesses will be sandboxed, limiting access to the system.

This method must be called before any web process has been created, as early as possible in your application. Calling it later is a fatal error.

This is only implemented on Linux and is a no-op otherwise.

Feature: v2_26

enabled

if true enable sandboxing

fn set_spell_checking_enabled(&self, enabled: bool)

Enable or disable the spell checking feature.

enabled

Value to be set

fn set_spell_checking_languages(&self, languages: &[&str])

Set the list of spell checking languages to be used for spell checking.

The locale string typically is in the form lang_COUNTRY, where lang is an ISO-639 language code, and COUNTRY is an ISO-3166 country code. For instance, sv_FI for Swedish as written in Finland or pt_BR for Portuguese as written in Brazil.

You need to call this function with a valid list of languages at least once in order to properly enable the spell checking feature in WebKit.

languages

a None-terminated list of spell checking languages

fn set_tls_errors_policy(&self, policy: TLSErrorsPolicy)

Set the TLS errors policy of self as policy

policy

a TLSErrorsPolicy

fn set_use_system_appearance_for_scrollbars(&self, enabled: bool)

Set the WebContext:use-system-appearance-for-scrollbars property.

Feature: v2_30

enabled

value to set

fn set_web_extensions_directory(&self, directory: &str)

Set the directory where WebKit will look for Web Extensions. This method must be called before loading anything in this context, otherwise it will not have any effect. You can connect to WebContext::initialize-web-extensions to call this method before anything is loaded.

directory

the directory to add

fn set_web_extensions_initialization_user_data(&self, user_data: &Variant)

Set user data to be passed to Web Extensions on initialization. The data will be passed to the WebKitWebExtensionInitializeWithUserDataFunction. This method must be called before loading anything in this context, otherwise it will not have any effect. You can connect to WebContext::initialize-web-extensions to call this method before anything is loaded.

Feature: v2_4

user_data

a glib::Variant

fn set_web_process_count_limit(&self, limit: u32)

👎 Deprecated

Sets the maximum number of web processes that can be created at the same time for the self. The default value is 0 and means no limit.

This function is now deprecated and does nothing for security reasons.

Feature: v2_10

Deprecated since 2.26

limit

the maximum number of web processes

fn get_property_local_storage_directory(&self) -> Option<GString>

👎 Deprecated

The directory where local storage data will be saved.

Feature: v2_8

Deprecated since 2.10

Use WebsiteDataManager:local-storage-directory instead.

fn get_property_process_swap_on_cross_site_navigation_enabled(&self) -> bool

Whether swap Web processes on cross-site navigations is enabled.

When enabled, pages from each security origin will be handled by their own separate Web processes, which are started (and terminated) on demand as the user navigates across different domains. This is an important security measure which helps prevent websites stealing data from other visited pages.

Feature: v2_28

fn connect_automation_started<F: Fn(&Self, &AutomationSession) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when a new automation request is made. Note that it will never be emitted if automation is not enabled in context, see WebContextExt::set_automation_allowed for more details.

Feature: v2_18

session

the AutomationSession associated with this event

fn connect_download_started<F: Fn(&Self, &Download) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when a new download request is made.

download

the Download associated with this event

fn connect_initialize_notification_permissions<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when a WebContext needs to set initial notification permissions for a web process. It is emitted when a new web process is about to be launched, and signals the most appropriate moment to use WebContextExt::initialize_notification_permissions. If no notification permissions have changed since the last time this signal was emitted, then there is no need to call WebContextExt::initialize_notification_permissions again.

Feature: v2_16

fn connect_initialize_web_extensions<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when a new web process is about to be launched. It signals the most appropriate moment to use WebContextExt::set_web_extensions_initialization_user_data and WebContextExt::set_web_extensions_directory.

Feature: v2_4

fn connect_user_message_received<F: Fn(&Self, &UserMessage) -> bool + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when a UserMessage is received from a WebKitWebExtension. You can reply to the message using UserMessageExt::send_reply.

You can handle the user message asynchronously by calling gobject::ObjectExt::ref on message and returning true.

Feature: v2_28

message

the UserMessage received

Returns

true if the message was handled, or false otherwise.

fn connect_property_use_system_appearance_for_scrollbars_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...