Trait webkit2gtk::WebResourceExt

pub trait WebResourceExt: 'static {
    fn get_data<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<u8>, Error>) + Send + 'static>(
        &self, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn get_data_future(
        &self
    ) -> Pin<Box_<dyn Future<Output = Result<Vec<u8>, Error>> + 'static>>;

    fn get_response(&self) -> Option<URIResponse>;

    fn get_uri(&self) -> Option<GString>;

    fn connect_failed<F: Fn(&Self, &Error) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_failed_with_tls_errors<F: Fn(&Self, &TlsCertificate, TlsCertificateFlags) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_finished<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn connect_received_data<F: Fn(&Self, u64) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_sent_request<F: Fn(&Self, &URIRequest, &URIResponse) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_response_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_uri_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all WebResource methods.

Implementors

WebResource

Required methods

fn get_data<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<u8>, Error>) + Send + 'static>(
&self,
cancellable: Option<&P>,
callback: Q
)

Asynchronously get the raw data for self.

When the operation is finished, callback will be called. You can then call WebResourceExt::get_data_finish to get the result of the operation.

cancellable

a gio::Cancellable or None to ignore

callback

a GAsyncReadyCallback to call when the request is satisfied

user_data

the data to pass to callback function

fn get_data_future(
&self
) -> Pin<Box_<dyn Future<Output = Result<Vec<u8>, Error>> + 'static>>

fn get_response(&self) -> Option<URIResponse>

Retrieves the URIResponse of the resource load operation. This method returns None if called before the response is received from the server. You can connect to notify::response signal to be notified when the response is received.

Returns

the URIResponse, or None if the response hasn't been received yet.

fn get_uri(&self) -> Option<GString>

Returns the current active URI of self. The active URI might change during a load operation:

<orderedlist> <listitem>``<para> When the resource load starts, the active URI is the requested URI </para>``</listitem> <listitem>``<para> When the initial request is sent to the server, WebResource::sent-request signal is emitted without a redirected response, the active URI is the URI of the request sent to the server. </para>``</listitem> <listitem>``<para> In case of a server redirection, WebResource::sent-request signal is emitted again with a redirected response, the active URI is the URI the request was redirected to. </para>``</listitem> <listitem>``<para> When the response is received from the server, the active URI is the final one and it will not change again. </para>``</listitem> </orderedlist>

You can monitor the active URI by connecting to the notify::uri signal of self.

Returns

the current active URI of self

fn connect_failed<F: Fn(&Self, &Error) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when an error occurs during the resource load operation.

error

the glib::Error that was triggered

fn connect_failed_with_tls_errors<F: Fn(&Self, &TlsCertificate, TlsCertificateFlags) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when a TLS error occurs during the resource load operation.

Feature: v2_8

certificate

a gio::TlsCertificate

errors

a gio::TlsCertificateFlags with the verification status of certificate

fn connect_finished<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted when the resource load finishes successfully or due to an error. In case of errors WebResource::failed signal is emitted before this one.

fn connect_received_data<F: Fn(&Self, u64) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted after response is received, every time new data has been received. It's useful to know the progress of the resource load operation.

data_length

the length of data received in bytes

fn connect_sent_request<F: Fn(&Self, &URIRequest, &URIResponse) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted when request has been sent to the server. In case of a server redirection this signal is emitted again with the request argument containing the new request sent to the server due to the redirection and the redirected_response parameter containing the response received by the server for the initial request.

request

a URIRequest

redirected_response

a URIResponse, or None

fn connect_property_response_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_uri_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

Loading content...