Trait webkit2gtk::WebsiteDataManagerExt

pub trait WebsiteDataManagerExt: 'static {
    fn clear<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
        &self, 
        types: WebsiteDataTypes, 
        timespan: TimeSpan, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn clear_future(
        &self, 
        types: WebsiteDataTypes, 
        timespan: TimeSpan
    ) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>;

    fn fetch<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<WebsiteData>, Error>) + Send + 'static>(
        &self, 
        types: WebsiteDataTypes, 
        cancellable: Option<&P>, 
        callback: Q
    );

    fn fetch_future(
        &self, 
        types: WebsiteDataTypes
    ) -> Pin<Box_<dyn Future<Output = Result<Vec<WebsiteData>, Error>> + 'static>>;

    fn get_base_cache_directory(&self) -> Option<GString>;

    fn get_base_data_directory(&self) -> Option<GString>;

    fn get_cookie_manager(&self) -> Option<CookieManager>;

    fn get_disk_cache_directory(&self) -> Option<GString>;

    fn get_hsts_cache_directory(&self) -> Option<GString>;

    fn get_indexeddb_directory(&self) -> Option<GString>;

    fn get_local_storage_directory(&self) -> Option<GString>;

    fn get_offline_application_cache_directory(&self) -> Option<GString>;

    fn get_websql_directory(&self) -> Option<GString>;

    fn is_ephemeral(&self) -> bool;

    fn get_property_is_ephemeral(&self) -> bool;
}

Trait containing all WebsiteDataManager methods.

Feature: v2_10

Implementors

WebsiteDataManager

Required methods

fn clear<P: IsA<Cancellable>, Q: FnOnce(Result<(), Error>) + Send + 'static>(
&self,
types: WebsiteDataTypes,
timespan: TimeSpan,
cancellable: Option<&P>,
callback: Q
)

Asynchronously clear the website data of the given types modified in the past timespan. If timespan is 0, all website data will be removed.

When the operation is finished, callback will be called. You can then call WebsiteDataManagerExt::clear_finish to get the result of the operation.

Due to implementation limitations, this function does not currently delete any stored cookies if timespan is nonzero. This behavior may change in the future.

Feature: v2_16

types

WebsiteDataTypes

timespan

a glib::TimeSpan

cancellable

a gio::Cancellable or None to ignore

callback

a GAsyncReadyCallback to call when the request is satisfied

user_data

the data to pass to callback function

fn clear_future(
&self,
types: WebsiteDataTypes,
timespan: TimeSpan
) -> Pin<Box_<dyn Future<Output = Result<(), Error>> + 'static>>

fn fetch<P: IsA<Cancellable>, Q: FnOnce(Result<Vec<WebsiteData>, Error>) + Send + 'static>(
&self,
types: WebsiteDataTypes,
cancellable: Option<&P>,
callback: Q
)

Asynchronously get the list of WebsiteData for the given types.

When the operation is finished, callback will be called. You can then call WebsiteDataManagerExt::fetch_finish to get the result of the operation.

Feature: v2_16

types

WebsiteDataTypes

cancellable

a gio::Cancellable or None to ignore

callback

a GAsyncReadyCallback to call when the request is satisfied

user_data

the data to pass to callback function

fn fetch_future(
&self,
types: WebsiteDataTypes
) -> Pin<Box_<dyn Future<Output = Result<Vec<WebsiteData>, Error>> + 'static>>

fn get_base_cache_directory(&self) -> Option<GString>

Get the WebsiteDataManager:base-cache-directory property.

Feature: v2_10

Returns

the base directory for Website cache, or None if WebsiteDataManager:base-cache-directory was not provided or self is ephemeral.

fn get_base_data_directory(&self) -> Option<GString>

Get the WebsiteDataManager:base-data-directory property.

Feature: v2_10

Returns

the base directory for Website data, or None if WebsiteDataManager:base-data-directory was not provided or self is ephemeral.

fn get_cookie_manager(&self) -> Option<CookieManager>

Get the CookieManager of self.

Feature: v2_16

Returns

a CookieManager

fn get_disk_cache_directory(&self) -> Option<GString>

Get the WebsiteDataManager:disk-cache-directory property.

Feature: v2_10

Returns

the directory where HTTP disk cache is stored or None if self is ephemeral.

fn get_hsts_cache_directory(&self) -> Option<GString>

Get the WebsiteDataManager:hsts-cache-directory property.

Feature: v2_26

Returns

the directory where the HSTS cache is stored or None if self is ephemeral.

fn get_indexeddb_directory(&self) -> Option<GString>

Get the WebsiteDataManager:indexeddb-directory property.

Feature: v2_10

Returns

the directory where IndexedDB databases are stored or None if self is ephemeral.

fn get_local_storage_directory(&self) -> Option<GString>

Get the WebsiteDataManager:local-storage-directory property.

Feature: v2_10

Returns

the directory where local storage data is stored or None if self is ephemeral.

fn get_offline_application_cache_directory(&self) -> Option<GString>

Get the WebsiteDataManager:offline-application-cache-directory property.

Feature: v2_10

Returns

the directory where offline web application cache is stored or None if self is ephemeral.

fn get_websql_directory(&self) -> Option<GString>

👎 Deprecated

Get the WebsiteDataManager:websql-directory property.

Feature: v2_10

Deprecated since 2.24

WebSQL is no longer supported. Use IndexedDB instead.

Returns

the directory where WebSQL databases are stored or None if self is ephemeral.

fn is_ephemeral(&self) -> bool

Get whether a WebsiteDataManager is ephemeral. See WebsiteDataManager:is-ephemeral for more details.

Feature: v2_16

Returns

true if self is ephemeral or false otherwise.

fn get_property_is_ephemeral(&self) -> bool

Whether the WebsiteDataManager is ephemeral. An ephemeral WebsiteDataManager handles all websites data as non-persistent, and nothing will be written to the client storage. Note that if you create an ephemeral WebsiteDataManager all other construction parameters to configure data directories will be ignored.

Feature: v2_16

Loading content...