Trait webkit2gtk::WindowPropertiesExt

pub trait WindowPropertiesExt: 'static {
    fn get_fullscreen(&self) -> bool;

    fn get_geometry(&self) -> Rectangle;

    fn get_locationbar_visible(&self) -> bool;

    fn get_menubar_visible(&self) -> bool;

    fn get_resizable(&self) -> bool;

    fn get_scrollbars_visible(&self) -> bool;

    fn get_statusbar_visible(&self) -> bool;

    fn get_toolbar_visible(&self) -> bool;
}

Trait containing all WindowProperties methods.

Implementors

WindowProperties

Required methods

fn get_fullscreen(&self) -> bool

Get whether the window should be shown in fullscreen state or not.

Returns

true if the window should be fullscreen or false otherwise.

fn get_geometry(&self) -> Rectangle

Get the geometry the window should have on the screen when shown.

geometry

return location for the window geometry

fn get_locationbar_visible(&self) -> bool

Get whether the window should have the locationbar visible or not.

Returns

true if locationbar should be visible or false otherwise.

fn get_menubar_visible(&self) -> bool

Get whether the window should have the menubar visible or not.

Returns

true if menubar should be visible or false otherwise.

fn get_resizable(&self) -> bool

Get whether the window should be resizable by the user or not.

Returns

true if the window should be resizable or false otherwise.

fn get_scrollbars_visible(&self) -> bool

Get whether the window should have the scrollbars visible or not.

Returns

true if scrollbars should be visible or false otherwise.

fn get_statusbar_visible(&self) -> bool

Get whether the window should have the statusbar visible or not.

Returns

true if statusbar should be visible or false otherwise.

fn get_toolbar_visible(&self) -> bool

Get whether the window should have the toolbar visible or not.

Returns

true if toolbar should be visible or false otherwise.

Loading content...