Crate webkit2gtk

webkit2gtk

Rust bindings and wrappers for webkit2gtk.

Building

webkit2gtk-rs expects GTK+, GLib and webkit2gtk development files to be installed on your system. See the requirements page.

Using

[dependencies]
webkit2gtk = "0.18.0"

Simple Example Browser

You can run the example in example/main.rs with

cargo run --example main

License

webkit2gtk-rs is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gio;

pub use glib;

Structs

ApplicationInfo

AuthenticationRequest

AutomationSession

AutomationSessionBuilder

A builder-pattern type to construct AutomationSession objects.

BackForwardList

BackForwardListItem

ColorChooserRequest

ColorChooserRequestBuilder

A builder-pattern type to construct ColorChooserRequest objects.

ContextMenu

ContextMenuItem

CookieManager

Credential

DeviceInfoPermissionRequest

Download

DownloadBuilder

A builder-pattern type to construct Download objects.

EditorState

EditorTypingAttributes

Error

A generic error capable of representing various error domains (types).

FaviconDatabase

FileChooserRequest

FindController

FindControllerBuilder

A builder-pattern type to construct FindController objects.

FindOptions

FormSubmissionRequest

GeolocationManager

GeolocationPermissionRequest

GeolocationPosition

HitTestResult

HitTestResultBuilder

A builder-pattern type to construct HitTestResult objects.

HitTestResultContext

ITPFirstParty

ITPThirdParty

InputHints

InputMethodContext

InputMethodUnderline

InstallMissingMediaPluginsPermissionRequest

JavascriptResult

MediaKeySystemPermissionRequest

MemoryPressureSettings

MimeInfo

NavigationAction

NavigationPolicyDecision

NetworkProxySettings

Notification

NotificationPermissionRequest

OptionMenu

OptionMenuItem

PermissionRequest

PluginDeprecated

PointerLockPermissionRequest

PolicyDecision

PrintCustomWidget

PrintCustomWidgetBuilder

A builder-pattern type to construct PrintCustomWidget objects.

PrintOperation

PrintOperationBuilder

A builder-pattern type to construct PrintOperation objects.

ResponsePolicyDecision

ScriptDialog

SecurityManager

SecurityOrigin

Settings

SettingsBuilder

A builder-pattern type to construct Settings objects.

SnapshotOptions

URIRequest

URIRequestBuilder

A builder-pattern type to construct URIRequest objects.

URIResponse

URISchemeRequest

URISchemeResponse

URISchemeResponseBuilder

A builder-pattern type to construct URISchemeResponse objects.

UserContentManager

UserMediaPermissionRequest

UserMessage

UserMessageBuilder

A builder-pattern type to construct UserMessage objects.

UserScript

UserStyleSheet

WebContext

WebContextBuilder

A builder-pattern type to construct WebContext objects.

WebInspector

WebResource

WebView

WebViewBase

WebViewBuilder

A builder-pattern type to construct WebView objects.

WebViewSessionState

WebsiteData

WebsiteDataAccessPermissionRequest

WebsiteDataManager

WebsiteDataManagerBuilder

A builder-pattern type to construct WebsiteDataManager objects.

WebsiteDataTypes

WebsitePolicies

WebsitePoliciesBuilder

A builder-pattern type to construct WebsitePolicies objects.

WindowProperties

WindowPropertiesBuilder

A builder-pattern type to construct WindowProperties objects.

Enums

AuthenticationScheme

AutomationBrowsingContextPresentation

AutoplayPolicy

CacheModel

ContextMenuAction

CookieAcceptPolicy

CookiePersistentStorage

CredentialPersistence

DownloadError

FaviconDatabaseError

HardwareAccelerationPolicy

InputPurpose

InsecureContentEvent

JavascriptError

LoadEvent

MediaCaptureState

NavigationType

NetworkError

NetworkProxyMode

PluginError

PolicyDecisionType

PolicyError

PrintError

PrintOperationResponse

ProcessModel

SaveMode

ScriptDialogType

SnapshotError

SnapshotRegion

TLSErrorsPolicy

UserContentFilterError

UserContentInjectedFrames

UserMessageError

UserScriptInjectionTime

UserStyleLevel

WebProcessTerminationReason

Traits

AuthenticationRequestExt

AutomationSessionExt

BackForwardListExt

BackForwardListItemExt

ColorChooserRequestExt

ContextMenuExt

ContextMenuItemExt

CookieManagerExt

DownloadExt

EditorStateExt

FaviconDatabaseExt

FileChooserRequestExt

FindControllerExt

FormSubmissionRequestExt

GeolocationManagerExt

HitTestResultExt

InputMethodContextExt

InstallMissingMediaPluginsPermissionRequestExt

NavigationPolicyDecisionExt

NotificationExt

OptionMenuExt

PermissionRequestExt

PluginExtDeprecated

PolicyDecisionExt

PrintCustomWidgetExt

PrintOperationExt

ResponsePolicyDecisionExt

SecurityManagerExt

SettingsExt

URIRequestExt

URIResponseExt

URISchemeRequestExt

URISchemeResponseExt

UserContentManagerExt

UserMediaPermissionRequestExt

UserMessageExt

WebContextExt

WebInspectorExt

WebResourceExt

WebViewExt

WebViewExtManual

WebsiteDataAccessPermissionRequestExt

WebsiteDataManagerExt

WebsiteDataManagerExtManual

WebsitePoliciesExt

WindowPropertiesExt