Crate webkit2gtk

Structs

ApplicationInfo

AuthenticationRequest

AuthenticationRequestClass

AutomationSession

AutomationSessionBuilder

AutomationSessionClass

BackForwardList

BackForwardListClass

BackForwardListItem

BackForwardListItemClass

ColorChooserRequest

ColorChooserRequestBuilder

ColorChooserRequestClass

ContextMenu

ContextMenuClass

ContextMenuItem

ContextMenuItemClass

CookieManager

CookieManagerClass

Credential

DeviceInfoPermissionRequest

DeviceInfoPermissionRequestClass

Download

DownloadBuilder

DownloadClass

EditorState

EditorStateClass

EditorTypingAttributes

Error
A generic error capable of representing various error domains (types).

FaviconDatabase

FaviconDatabaseClass

FileChooserRequest

FileChooserRequestClass

FindController

FindControllerBuilder

FindControllerClass

FindOptions

FormSubmissionRequest

FormSubmissionRequestClass

GeolocationManager

GeolocationManagerClass

GeolocationPermissionRequest

GeolocationPermissionRequestClass

GeolocationPosition

HitTestResult

HitTestResultBuilder

HitTestResultClass

HitTestResultContext

InputHints

InputMethodContext

InputMethodContextClass

InputMethodUnderline

InstallMissingMediaPluginsPermissionRequest

InstallMissingMediaPluginsPermissionRequestClass

JavascriptResult

MimeInfo

NavigationAction

NavigationPolicyDecision

NavigationPolicyDecisionClass

NetworkProxySettings

Notification

NotificationClass

NotificationPermissionRequest

NotificationPermissionRequestClass

OptionMenu

OptionMenuClass

OptionMenuItem

PermissionRequest

Plugin

PluginClass

PointerLockPermissionRequest

PointerLockPermissionRequestClass

PolicyDecision

PolicyDecisionClass

PrintCustomWidget

PrintCustomWidgetBuilder

PrintCustomWidgetClass

PrintOperation

PrintOperationBuilder

PrintOperationClass

ResponsePolicyDecision

ResponsePolicyDecisionClass

ScriptDialog

SecurityManager

SecurityManagerClass

SecurityOrigin

Settings

SettingsBuilder

SettingsClass

SnapshotOptions

URIRequest

URIRequestBuilder

URIRequestClass

URIResponse

URIResponseClass

URISchemeRequest

URISchemeRequestClass

UserContentManager

UserContentManagerClass

UserMediaPermissionRequest

UserMediaPermissionRequestClass

UserMessage

UserMessageBuilder

UserMessageClass

UserScript

UserStyleSheet

WebContext

WebContextBuilder

WebContextClass

WebInspector

WebInspectorClass

WebResource

WebResourceClass

WebView

WebViewBase

WebViewBaseClass

WebViewBuilder

WebViewClass

WebViewSessionState

WebsiteData

WebsiteDataManager

WebsiteDataManagerBuilder

WebsiteDataManagerClass

WebsiteDataTypes

WindowProperties

WindowPropertiesBuilder

WindowPropertiesClass

Enums

AuthenticationScheme
Enum values representing the authentication scheme.

AutomationBrowsingContextPresentation
Enum values used for determining the automation browsing context presentation.

CacheModel
Enum values used for determining the WebContext cache model.

ContextMenuAction
Enum values used to denote the stock actions for ContextMenuItems

CookieAcceptPolicy
Enum values used to denote the cookie acceptance policies.

CookiePersistentStorage
Enum values used to denote the cookie persistent storage types.

CredentialPersistence
Enum values representing the duration for which a credential persists.

DownloadError
Enum values used to denote the various download errors.

FaviconDatabaseError
Enum values used to denote the various errors related to the FaviconDatabase.

HardwareAccelerationPolicy
Enum values used for determining the hardware acceleration policy.

InputPurpose
Enum values used to describe the primary purpose of the active editable element.

InsecureContentEvent
Enum values used to denote the different events which can trigger the detection of insecure content.

JavascriptError
Enum values used to denote errors happening when executing JavaScript

LoadEvent
Enum values used to denote the different events that happen during a WebView load operation.

NavigationType
Enum values used to denote the various navigation types.

NetworkError
Enum values used to denote the various network errors.

NetworkProxyMode
Enum values used to set the network proxy mode.

PluginError
Enum values used to denote the various plugin errors.

PolicyDecisionType
Enum values used for determining the type of a policy decision during WebView::decide-policy.

PolicyError
Enum values used to denote the various policy errors.

PrintError
Enum values used to denote the various print errors.

PrintOperationResponse
Enum values representing the response of the print dialog shown with PrintOperationExt::run_dialog.

ProcessModel
Enum values used for determining the WebContext process model.

SaveMode
Enum values to specify the different ways in which a WebView can save its current web page into a self-contained file.

ScriptDialogType
Enum values used for determining the type of ScriptDialog

SnapshotError
Enum values used to denote errors happening when creating snapshots of WebView

SnapshotRegion
Enum values used to specify the region from which to get a WebView snapshot

TLSErrorsPolicy
Enum values used to denote the TLS errors policy.

UserContentFilterError

UserContentInjectedFrames
Specifies in which frames user style sheets are to be inserted in.

UserMessageError
Enum values used to denote errors happening when sending user messages.

UserScriptInjectionTime
Specifies at which place of documents an user script will be inserted.

UserStyleLevel
Specifies how to treat an user style sheet.

WebProcessTerminationReason
Enum values used to specify the reason why the web process terminated abnormally.

Constants

NONE_AUTHENTICATION_REQUEST

NONE_AUTOMATION_SESSION

NONE_BACK_FORWARD_LIST

NONE_BACK_FORWARD_LIST_ITEM

NONE_COLOR_CHOOSER_REQUEST

NONE_CONTEXT_MENU

NONE_CONTEXT_MENU_ITEM

NONE_COOKIE_MANAGER

NONE_DEVICE_INFO_PERMISSION_REQUEST

NONE_DOWNLOAD

NONE_EDITOR_STATE

NONE_FAVICON_DATABASE

NONE_FILE_CHOOSER_REQUEST

NONE_FIND_CONTROLLER

NONE_FORM_SUBMISSION_REQUEST

NONE_GEOLOCATION_MANAGER

NONE_GEOLOCATION_PERMISSION_REQUEST

NONE_HIT_TEST_RESULT

NONE_INPUT_METHOD_CONTEXT

NONE_INSTALL_MISSING_MEDIA_PLUGINS_PERMISSION_REQUEST

NONE_NAVIGATION_POLICY_DECISION

NONE_NOTIFICATION

NONE_NOTIFICATION_PERMISSION_REQUEST

NONE_OPTION_MENU

NONE_PERMISSION_REQUEST

NONE_PLUGIN

NONE_POINTER_LOCK_PERMISSION_REQUEST

NONE_POLICY_DECISION

NONE_PRINT_CUSTOM_WIDGET

NONE_PRINT_OPERATION

NONE_RESPONSE_POLICY_DECISION

NONE_SECURITY_MANAGER

NONE_SETTINGS

NONE_URI_REQUEST

NONE_URI_RESPONSE

NONE_URI_SCHEME_REQUEST

NONE_USER_CONTENT_MANAGER

NONE_USER_MEDIA_PERMISSION_REQUEST

NONE_USER_MESSAGE

NONE_WEBSITE_DATA_MANAGER

NONE_WEB_CONTEXT

NONE_WEB_INSPECTOR

NONE_WEB_RESOURCE

NONE_WEB_VIEW

NONE_WEB_VIEW_BASE

NONE_WINDOW_PROPERTIES

Traits

AuthenticationRequestExt
Trait containing all AuthenticationRequest methods.

AutomationSessionExt
Trait containing all AutomationSession methods.

BackForwardListExt
Trait containing all BackForwardList methods.

BackForwardListItemExt
Trait containing all BackForwardListItem methods.

ColorChooserRequestExt
Trait containing all ColorChooserRequest methods.

ContextMenuExt
Trait containing all ContextMenu methods.

ContextMenuItemExt
Trait containing all ContextMenuItem methods.

CookieManagerExt
Trait containing all CookieManager methods.

DownloadExt
Trait containing all Download methods.

EditorStateExt
Trait containing all EditorState methods.

FaviconDatabaseExt
Trait containing all FaviconDatabase methods.

FileChooserRequestExt
Trait containing all FileChooserRequest methods.

FindControllerExt
Trait containing all FindController methods.

FormSubmissionRequestExt
Trait containing all FormSubmissionRequest methods.

GeolocationManagerExt
Trait containing all GeolocationManager methods.

HitTestResultExt
Trait containing all HitTestResult methods.

InputMethodContextExt
Trait containing all InputMethodContext methods.

InstallMissingMediaPluginsPermissionRequestExt
Trait containing all InstallMissingMediaPluginsPermissionRequest methods.

NavigationPolicyDecisionExt
Trait containing all NavigationPolicyDecision methods.

NotificationExt
Trait containing all Notification methods.

OptionMenuExt
Trait containing all OptionMenu methods.

PermissionRequestExt
Trait containing all PermissionRequest methods.

PluginExt
Trait containing all Plugin methods.

PolicyDecisionExt
Trait containing all PolicyDecision methods.

PrintCustomWidgetExt
Trait containing all PrintCustomWidget methods.

PrintOperationExt
Trait containing all PrintOperation methods.

ResponsePolicyDecisionExt
Trait containing all ResponsePolicyDecision methods.

SecurityManagerExt
Trait containing all SecurityManager methods.

SettingsExt
Trait containing all Settings methods.

URIRequestExt
Trait containing all URIRequest methods.

URIResponseExt
Trait containing all URIResponse methods.

URISchemeRequestExt
Trait containing all URISchemeRequest methods.

UserContentManagerExt
Trait containing all UserContentManager methods.

UserMediaPermissionRequestExt
Trait containing all UserMediaPermissionRequest methods.

UserMessageExt
Trait containing all UserMessage methods.

WebContextExt
Trait containing all WebContext methods.

WebInspectorExt
Trait containing all WebInspector methods.

WebResourceExt
Trait containing all WebResource methods.

WebViewExt
Trait containing all WebView methods.

WebsiteDataManagerExt
Trait containing all WebsiteDataManager methods.

WindowPropertiesExt
Trait containing all WindowProperties methods.